refactor: split decision and fact modules
