+++ packages/daemon/src/protocol/daemon-protocol.contract.ts
@@ -118,7 +118,7 @@ export type DaemonGuiReadResultMap = {
   readonly "repo.tasks.list": DaemonTaskSnapshotListResult;
   readonly "repo.agenda.read": DaemonAgendaResult;
   readonly "repo.triadic.relationGraph": { readonly ok: true } & Omit<ReturnType<typeof import("../../../kernel/src/projection/sqlite-task-projection.ts").readRelationGraphProjection>, "taskRows"> & { readonly page?: ProjectionPage };
-  readonly "repo.decisions.list": { readonly ok: true; readonly decisions: readonly import("../../../kernel/src/projection/fact-event-projection.ts").DecisionProjectionRow[]; readonly warnings: readonly import("../../../kernel/src/projection/types.ts").ProjectionWarning[] }; readonly "repo.tasks.document.read": { readonly ok: true; readonly status: "ready" | "pending"; readonly taskId: string; readonly path: string; readonly body: string; readonly blobSha256: string | null; readonly watermark: number; readonly sourceRevision: number }; readonly "repo.tasks.documents.list": DaemonTaskDocumentListResult;
+  readonly "repo.decisions.list": { readonly ok: true; readonly decisions: readonly import("../../../kernel/src/projection/decision-event-projection.ts").DecisionProjectionRow[]; readonly warnings: readonly import("../../../kernel/src/projection/types.ts").ProjectionWarning[] }; readonly "repo.tasks.document.read": { readonly ok: true; readonly status: "ready" | "pending"; readonly taskId: string; readonly path: string; readonly body: string; readonly blobSha256: string | null; readonly watermark: number; readonly sourceRevision: number }; readonly "repo.tasks.documents.list": DaemonTaskDocumentListResult;
   readonly "repo.agentRuntime.overview": AgentRuntimeOverviewResult; readonly "repo.agentRuntime.sessions.read": AgentRuntimeSessionResult; readonly "repo.agentRuntime.events.read": AgentRuntimeEventsResult; readonly "repo.task.dispatches": DaemonTaskDispatchesResult;
   readonly "repo.agent.entities.list": Extract<AgentEntityGuiRead, { readonly schema: "agent-entity-catalog/v1" }>; readonly "repo.agent.entity.read": Extract<AgentEntityGuiRead, { readonly schema: "agent-entity-detail/v1" }>; readonly "repo.squad.entities.list": Extract<AgentEntityGuiRead, { readonly schema: "squad-entity-catalog/v1" }>; readonly "repo.squad.entity.read": Extract<AgentEntityGuiRead, { readonly schema: "squad-entity-detail/v1" }>;
   readonly "repo.gui.catalog.snapshot": JsonObject; readonly "repo.gui.catalog.preset.read": JsonObject; readonly "repo.terminal.sessions.list": JsonObject;
