+++ packages/kernel/src/domain/decision-event.ts
@@ -1,0 +1,114 @@
+import { sha256Text, stableStringify } from "../integrity/stable-hash.ts";
+import { normalizeRelativeDocumentPath } from "../layout/portable-path.ts";
+import { eventObjectTarget } from "../layout/ledger-object-layout.ts";
+import { freezeDeclaredWritePlan, hasContractFields as matchesFields, isFrozenWritePlan, isNonEmptyString, isRecord, sameActorIdentity, sameWriteSource, serializeEventEnvelope, validateActorIdentity, validateEventEnvelopeIdentity, validateWriteSource, type ActorIdentity, type EventEnvelope, type FrozenWritePlan, type WriteTarget } from "./write-chain.contract.ts";
+import { deriveRelationId, relationDirections, relationOrigins, relationStates, relationStrengths, relationTypes, type EntityRelationRecord } from "./entity-relation.ts";
+import { codePoints, requiredWithOptional } from "./event-validation.ts";
+import { timestamp } from "./timestamp.ts";
+
+export const decisionEventTypes = ["decision_proposed", "decision_accepted", "decision_rejected", "decision_deferred", "decision_superseded", "decision_retired", "decision_amended", "decision_repinned", "decision_claim_declared", "decision_claim_fulfillment_declared", "decision_related", "decision_relation_retired", "decision_relation_replaced"] as const;
+export const decisionStates = ["proposed", "in_effect", "rejected", "deferred", "superseded", "outcome_retired"] as const;
+export const decisionFulfillmentModes = ["evidenced", "delivered", "standing_policy"] as const;
+export type DecisionState = typeof decisionStates[number]; export type DecisionFulfillmentMode = typeof decisionFulfillmentModes[number];
+export interface DecisionProposalPayload { readonly title: string; readonly question: string; readonly riskTier: "low" | "medium" | "high"; readonly urgency: "low" | "medium" | "high"; readonly vertical: string; readonly preset: string; readonly appliesTo: { readonly modules: readonly string[]; readonly productLines: readonly string[] }; readonly decisionClass: "ordinary" | "standing_policy"; readonly chosen: readonly { readonly id: string; readonly text: string; readonly rationale?: string }[]; readonly rejected: readonly { readonly id: string; readonly text: string; readonly whyNot: string }[]; readonly body: string; readonly claims: readonly { readonly id: string; readonly text: string; readonly loadBearing: boolean }[]; readonly fulfillments: readonly { readonly claimId: string; readonly mode: DecisionFulfillmentMode }[]; readonly relations: readonly EntityRelationRecord[] }
+export interface DecisionAmendableSnapshot { readonly title: string; readonly decisionClass: "ordinary" | "standing_policy"; readonly chosen: DecisionProposalPayload["chosen"]; readonly rejected: DecisionProposalPayload["rejected"]; readonly claims: readonly { readonly id: string; readonly text: string; readonly loadBearing: boolean; readonly fulfillment: DecisionFulfillmentMode | null }[] }
+export interface DecisionAmendmentV1 { readonly schema: "decision-amendment/v1"; readonly amendmentId: string; readonly fields: readonly string[]; readonly actor: ActorIdentity; readonly amendedAt: string }
+export interface DecisionContentPinV1 { readonly schema: "decision-content-pin/v1"; readonly pinId: string; readonly action: "accept" | "reject" | "defer" | "supersede" | "retire" | "amend" | "repin"; readonly state: DecisionState; readonly pinnedAt: string; readonly evidence: string; readonly actor: ActorIdentity; readonly digest: `sha256:${string}` }
+export interface DecisionPayloads { readonly decision_proposed: DecisionProposalPayload; readonly decision_accepted: { readonly rationale: string; readonly judgmentOnlyRationale: string | null; readonly fulfillments?: readonly { readonly claimId: string; readonly mode: DecisionFulfillmentMode }[]; readonly standingPolicy?: boolean }; readonly decision_rejected: { readonly reason: string }; readonly decision_deferred: { readonly reason: string }; readonly decision_superseded: { readonly reason: string }; readonly decision_retired: { readonly reason: string }; readonly decision_amended: { readonly next: DecisionAmendableSnapshot; readonly fields: readonly string[]; readonly body: string | null }; readonly decision_repinned: { readonly migrationEvidence: string }; readonly decision_claim_declared: { readonly claimId: string; readonly text: string; readonly loadBearing: boolean }; readonly decision_claim_fulfillment_declared: { readonly claimId: string; readonly mode: DecisionFulfillmentMode }; readonly decision_related: { readonly relation: EntityRelationRecord }; readonly decision_relation_retired: { readonly relationId: string; readonly reason: string }; readonly decision_relation_replaced: { readonly relationId: string; readonly reason: string; readonly replacement: EntityRelationRecord; readonly body: string | null } }
+export type DecisionJudgmentAction = "accept" | "reject" | "defer";
+export interface DecisionJudgmentConsentV1 { readonly schema: "decision-judgment-consent/v1"; readonly consentId: string; readonly decisionId: string; readonly action: DecisionJudgmentAction; readonly targetState: "in_effect" | "rejected" | "deferred"; readonly machineDigest: `sha256:${string}`; readonly actor: ActorIdentity; readonly source: DecisionEventDraftV1["source"]; readonly consentedAt: string }
+export const DECISION_DOCUMENT_POLICY_ID = "markdown-body-replaceable/v1" as const;
+export interface DecisionDocumentClaim { readonly path: string; readonly sha256: string; readonly size: number; readonly mediaType: "text/markdown"; readonly policyId: typeof DECISION_DOCUMENT_POLICY_ID }
+export interface DecisionDocumentMutation { readonly baseDocumentSha256: string | null; readonly decisionDocumentClaim: DecisionDocumentClaim }
+export type DecisionEventDraftV1 = { [T in keyof DecisionPayloads]: EventEnvelope<"decision-event/v1", T, ActorIdentity, DecisionPayloads[T]> & { readonly decisionId: string } }[keyof DecisionPayloads];
+type DecisionOutcomeType = "decision_accepted" | "decision_rejected" | "decision_deferred"; type DecisionTransitionType = DecisionOutcomeType | "decision_superseded" | "decision_retired";
+type DecisionPublishedPayload<T extends keyof DecisionPayloads> = DecisionPayloads[T] & DecisionDocumentMutation & (T extends DecisionOutcomeType ? { readonly judgmentConsent: DecisionJudgmentConsentV1 } : object) & (T extends DecisionTransitionType | "decision_repinned" ? { readonly contentPin?: DecisionContentPinV1 } : T extends "decision_amended" ? { readonly contentPin: DecisionContentPinV1 } : object) & (T extends "decision_amended" ? { readonly amendment: DecisionAmendmentV1 } : object);
+export type DecisionEventV1 = { [T in keyof DecisionPayloads]: EventEnvelope<"decision-event/v1", T, ActorIdentity, DecisionPublishedPayload<T>> & { readonly decisionId: string } }[keyof DecisionPayloads];
+export interface DecisionDocumentState { readonly decisionId: string; readonly state: DecisionState; readonly title: string; readonly question: string; readonly riskTier: "low" | "medium" | "high"; readonly urgency: "low" | "medium" | "high"; readonly vertical: string; readonly preset: string; readonly decisionClass: "ordinary" | "standing_policy"; readonly appliesTo: { readonly modules: readonly string[]; readonly productLines: readonly string[] }; readonly proposer: ActorIdentity; readonly arbiter: ActorIdentity | null; readonly proposedAt: string; readonly decidedAt: string | null; readonly workspaceRevision: number; readonly chosen: DecisionProposalPayload["chosen"]; readonly rejected: DecisionProposalPayload["rejected"]; readonly claims: readonly { readonly id: string; readonly text: string; readonly loadBearing: boolean; readonly fulfillment: DecisionFulfillmentMode | null }[]; readonly relations: readonly EntityRelationRecord[]; readonly judgmentConsents: readonly DecisionJudgmentConsentV1[]; readonly amendments?: readonly DecisionAmendmentV1[]; readonly contentPins?: readonly DecisionContentPinV1[] }
+export interface DecisionContentBlob { readonly sha256: string; readonly size: number; readonly mediaType: "text/markdown"; readonly body: string }
+export interface CompiledDecisionWrite { readonly event: DecisionEventV1; readonly plan: FrozenWritePlan<"DecisionWrite">; readonly blobs: readonly [DecisionContentBlob]; readonly path: string; readonly body: string }
+
+export function compileDecisionWrite(input: { readonly event: DecisionEventDraftV1; readonly currentDecision: Omit<DecisionDocumentState, "relations"> | null; readonly currentRelations: readonly EntityRelationRecord[]; readonly currentDocument: { readonly blobSha256: string; readonly body: string } | null }): CompiledDecisionWrite { const proposal = input.event.type === "decision_proposed"; if (proposal ? input.currentDecision !== null || input.currentDocument !== null : input.currentDecision === null || input.currentDocument === null) throw new Error("decision projection and authored document base must agree"); const base = input.currentDecision === null ? null : { ...input.currentDecision, relations: input.currentRelations }; assertDecisionEvidenceFloor(base, input.event); const reduced = reduceDecisionDocument(base, input.event), consent = base && decisionOutcome(input.event) ? decisionConsent(base, input.event) : null, amendment = base && input.event.type === "decision_amended" ? decisionAmendment(input.event) : null, pin = base && (decisionTransition(input.event) || input.event.type === "decision_amended" || input.event.type === "decision_repinned") ? decisionContentPin(reduced, input.event) : null, next = { ...reduced, judgmentConsents: consent ? [...reduced.judgmentConsents, consent] : reduced.judgmentConsents, amendments: amendment ? [...(reduced.amendments ?? []), amendment] : reduced.amendments, contentPins: pin ? [...(reduced.contentPins ?? []), pin] : reduced.contentPins }, path = `decisions/decision-${input.event.decisionId}/decision.md`; try { if (normalizeRelativeDocumentPath(path) !== path) throw new Error(); } catch { throw new Error("decision package path is invalid"); } const replacementBody = proposal ? input.event.payload.body : input.event.type === "decision_amended" || input.event.type === "decision_relation_replaced" ? input.event.payload.body ?? undefined : undefined, judgment = input.event.type === "decision_accepted" ? input.event.payload.judgmentOnlyRationale : null, body = renderDecisionDocument(next, input.currentDocument?.body ?? null, replacementBody, judgment), claim: DecisionDocumentClaim = { path, sha256: sha256Text(body), size: Buffer.byteLength(body), mediaType: "text/markdown", policyId: DECISION_DOCUMENT_POLICY_ID }, event = { ...input.event, payload: { ...input.event.payload, ...(consent ? { judgmentConsent: consent } : {}), ...(amendment ? { amendment } : {}), ...(pin ? { contentPin: pin } : {}), baseDocumentSha256: input.currentDocument?.blobSha256 ?? null, decisionDocumentClaim: claim } } as DecisionEventV1; return { event, plan: decisionWritePlan(event), blobs: [{ sha256: claim.sha256, size: claim.size, mediaType: claim.mediaType, body }], path, body }; }
+export function decisionWritePlan(event: DecisionEventV1): FrozenWritePlan<"DecisionWrite"> { const claim = event.payload.decisionDocumentClaim, targets: WriteTarget[] = [{ kind: "event_file", path: eventObjectTarget(event.opId), operation: "create" }, { kind: "event_head", path: "harness/events/head.json", operation: "replace" }, { kind: "authored_file", path: claim.path, operation: "replace", sha256: claim.sha256, size: claim.size, mediaType: claim.mediaType }, { kind: "content_blob", sha256: claim.sha256, size: claim.size, mediaType: claim.mediaType }, { kind: "projection_invalidation", projection: "decision/v1", key: event.decisionId }, { kind: "projection_invalidation", projection: "document/v1", key: claim.path }]; if (!["decision_rejected", "decision_deferred", "decision_repinned"].includes(event.type)) targets.push({ kind: "projection_invalidation", projection: "relation-graph/v1", key: `decision/${event.decisionId}` }); return freezeDeclaredWritePlan({ commandType: "DecisionWrite", targets }, ["DecisionWrite"]); }
+export function assertDecisionWritePlan(event: DecisionEventV1, plan: FrozenWritePlan | undefined): void { const shape = (value: FrozenWritePlan) => stableStringify({ commandType: value.commandType, targets: value.targets.map(stableStringify).sort() }); if (!plan || !isFrozenWritePlan(plan) || shape(plan) !== shape(decisionWritePlan(event))) throw new Error("decision write plan must exactly declare event, document, blob, and projections"); }
+export function renderDecisionDocument(value: DecisionDocumentState, current: string | null, replacementBody?: string, judgmentOnlyRationale: string | null = null): string { const baseProse = replacementBody ?? (current === null ? `\n# ${value.title}\n` : decisionDocumentProse(current)), prose = judgmentOnlyRationale ? `${baseProse.replace(/\s*$/u, "")}\n\n## Judgment-only acceptance\n\n${judgmentOnlyRationale}\n` : baseProse, history = [...(value.amendments === undefined ? [] : [`amendments: ${stableStringify(value.amendments)}`]), ...(value.contentPins === undefined ? [] : [`contentPins: ${stableStringify(value.contentPins)}`])], frontmatter = ["---", "schema: decision-package/v1", `decision_id: ${value.decisionId}`, `workspaceRevision: ${value.workspaceRevision}`, `title: ${stableStringify(value.title)}`, `state: ${value.state}`, `riskTier: ${value.riskTier}`, `urgency: ${value.urgency}`, `vertical: ${stableStringify(value.vertical)}`, `preset: ${stableStringify(value.preset)}`, `decisionClass: ${value.decisionClass}`, `applies_to: ${stableStringify(value.appliesTo)}`, `proposer: ${stableStringify(value.proposer)}`, `arbiter: ${stableStringify(value.arbiter)}`, `proposedAt: ${stableStringify(value.proposedAt)}`, `decidedAt: ${stableStringify(value.decidedAt)}`, `question: ${stableStringify(value.question)}`, `chosen: ${stableStringify(value.chosen)}`, `rejected: ${stableStringify(value.rejected)}`, `claims: ${stableStringify(value.claims)}`, `relations: ${stableStringify(value.relations)}`, `judgmentConsents: ${stableStringify(value.judgmentConsents)}`, ...history, "---"].join("\n"); return `${frontmatter}\n${prose}`; }
+export function decisionDocumentProse(body: string): string { const match = /^---\n[\s\S]*?\n---\n/u.exec(body); if (!match) throw new Error("decision document frontmatter is invalid"); return body.slice(match[0].length); }
+function reduceDecisionDocument(current: DecisionDocumentState | null, event: DecisionEventDraftV1): DecisionDocumentState { if (event.type === "decision_proposed") { const p = event.payload; return { decisionId: event.decisionId, state: "proposed", title: p.title, question: p.question, riskTier: p.riskTier, urgency: p.urgency, vertical: p.vertical, preset: p.preset, decisionClass: p.decisionClass, appliesTo: p.appliesTo, proposer: event.actor, arbiter: null, proposedAt: event.occurredAt, decidedAt: null, workspaceRevision: event.workspaceRevision, chosen: p.chosen, rejected: p.rejected, claims: p.claims.map((claim) => ({ ...claim, fulfillment: p.fulfillments.find((entry) => entry.claimId === claim.id)?.mode ?? null })), relations: [...p.relations].sort((a, b) => a.relation_id.localeCompare(b.relation_id)), judgmentConsents: [] }; } if (!current) throw new Error(`Decision ${event.decisionId} does not exist.`); const revision = { ...current, workspaceRevision: event.workspaceRevision }; if (decisionOutcome(event)) { const fulfillments = event.type === "decision_accepted" ? event.payload.fulfillments ?? [] : [], claims = event.type === "decision_accepted" ? current.claims.map((claim) => ({ ...claim, fulfillment: fulfillments.find((entry) => entry.claimId === claim.id)?.mode ?? claim.fulfillment })) : current.claims; return { ...revision, state: outcomeState(event.type), decisionClass: event.type === "decision_accepted" && event.payload.standingPolicy ? "standing_policy" : current.decisionClass, claims, arbiter: event.actor, decidedAt: event.occurredAt }; } if (event.type === "decision_superseded" || event.type === "decision_retired") return { ...revision, state: event.type === "decision_superseded" ? "superseded" : "outcome_retired", decidedAt: event.occurredAt }; if (event.type === "decision_amended") return { ...revision, ...event.payload.next }; if (event.type === "decision_repinned") return revision; if (event.type === "decision_claim_declared") return { ...revision, claims: [...current.claims, { id: event.payload.claimId, text: event.payload.text, loadBearing: event.payload.loadBearing, fulfillment: null }] }; if (event.type === "decision_claim_fulfillment_declared") return { ...revision, claims: current.claims.map((claim) => claim.id === event.payload.claimId ? { ...claim, fulfillment: event.payload.mode } : claim) }; if (event.type === "decision_related") return { ...revision, relations: [...current.relations, event.payload.relation].sort((a, b) => a.relation_id.localeCompare(b.relation_id)) }; if (event.type === "decision_relation_replaced") return { ...revision, relations: [...current.relations.map((relation) => relation.relation_id === event.payload.relationId ? { ...relation, state: "edge_retired" as const } : relation), event.payload.replacement].sort((a, b) => a.relation_id.localeCompare(b.relation_id)) }; return { ...revision, relations: current.relations.map((relation) => relation.relation_id === event.payload.relationId ? { ...relation, state: "edge_retired" } : relation) }; }
+export function decisionMachineDigest(value: DecisionDocumentState): `sha256:${string}` { const semantic = { schema: "decision-machine-content/v1", decisionId: value.decisionId, title: value.title, question: value.question, riskTier: value.riskTier, urgency: value.urgency, vertical: value.vertical, preset: value.preset, decisionClass: value.decisionClass, appliesTo: value.appliesTo, chosen: value.chosen, rejected: value.rejected, claims: value.claims, relations: value.relations }; return `sha256:${sha256Text(stableStringify(semantic))}`; }
+export function assertDecisionJudgmentConsent(current: DecisionDocumentState, event: DecisionEventV1): void { if (!decisionOutcome(event)) return; const expected = decisionConsent(current, event); if (stableStringify(event.payload.judgmentConsent) !== stableStringify(expected)) invalidDecision("Decision judgment consent does not match the machine content cut or event authority."); assertDecisionEvidenceFloor(current, event); }
+export function assertDecisionContentPin(current: DecisionDocumentState, event: DecisionEventV1): void { if ((!decisionTransition(event) && event.type !== "decision_amended" && event.type !== "decision_repinned") || event.payload.contentPin === undefined) return; const reduced = reduceDecisionDocument(current, event), expected = decisionContentPin(reduced, event); if (stableStringify(event.payload.contentPin) !== stableStringify(expected)) invalidDecision("Decision content pin does not match the projected machine content cut."); }
+function decisionConsent(current: DecisionDocumentState, event: Extract<DecisionEventDraftV1 | DecisionEventV1, { readonly type: DecisionOutcomeType }>): DecisionJudgmentConsentV1 { const action = event.type.slice("decision_".length, -2) as DecisionJudgmentAction; return { schema: "decision-judgment-consent/v1", consentId: `djc_${sha256Text(event.opId).slice(0, 26)}`, decisionId: event.decisionId, action, targetState: outcomeState(event.type), machineDigest: decisionMachineDigest(current), actor: event.actor, source: event.source, consentedAt: event.occurredAt }; }
+function decisionAmendment(event: Extract<DecisionEventDraftV1 | DecisionEventV1, { readonly type: "decision_amended" }>): DecisionAmendmentV1 { return { schema: "decision-amendment/v1", amendmentId: `dam_${sha256Text(event.opId).slice(0, 26)}`, fields: event.payload.fields, actor: event.actor, amendedAt: event.occurredAt }; }
+function decisionContentPin(current: DecisionDocumentState, event: Extract<DecisionEventDraftV1 | DecisionEventV1, { readonly type: DecisionTransitionType | "decision_amended" | "decision_repinned" }>): DecisionContentPinV1 { const action = event.type === "decision_repinned" ? "repin" : event.type === "decision_amended" ? "amend" : ({ decision_accepted: "accept", decision_rejected: "reject", decision_deferred: "defer", decision_superseded: "supersede", decision_retired: "retire" } as const)[event.type], evidence = event.type === "decision_repinned" ? event.payload.migrationEvidence : event.type === "decision_amended" ? `fields:${event.payload.fields.join(",")}` : event.type === "decision_accepted" ? event.payload.rationale : event.type === "decision_rejected" || event.type === "decision_deferred" || event.type === "decision_superseded" || event.type === "decision_retired" ? event.payload.reason : "transition"; return { schema: "decision-content-pin/v1", pinId: `dcp_${sha256Text(event.opId).slice(0, 26)}`, action, state: current.state, pinnedAt: event.occurredAt, evidence, actor: event.actor, digest: decisionMachineDigest(current) }; }
+function decisionOutcome(event: DecisionEventDraftV1 | DecisionEventV1): event is Extract<DecisionEventDraftV1 | DecisionEventV1, { readonly type: DecisionOutcomeType }> { return event.type === "decision_accepted" || event.type === "decision_rejected" || event.type === "decision_deferred"; }
+function decisionTransition(event: DecisionEventDraftV1 | DecisionEventV1): event is Extract<DecisionEventDraftV1 | DecisionEventV1, { readonly type: DecisionTransitionType }> { return decisionOutcome(event) || event.type === "decision_superseded" || event.type === "decision_retired"; }
+function outcomeState(type: DecisionOutcomeType): "in_effect" | "rejected" | "deferred" { return type === "decision_accepted" ? "in_effect" : type.slice("decision_".length) as "rejected" | "deferred"; }
+function assertDecisionEvidenceFloor(current: DecisionDocumentState | null, event: DecisionEventDraftV1 | DecisionEventV1): void { if (event.type !== "decision_accepted" || event.payload.judgmentOnlyRationale?.trim()) return; const claims = new Set((current?.claims ?? []).map((claim) => `decision/${event.decisionId}/${claim.id}`)), evidence = current?.relations.some((edge) => edge.state === "active" && claims.has(edge.source) && /^(?:fact\/[^/]+\/F-[0-9A-HJKMNP-TV-Z]{8}|task\/[^/]+|decision\/dec_[A-Za-z0-9_-]+(?:\/[A-Za-z][A-Za-z0-9_-]*)?)$/u.test(edge.target)); if (!evidence) invalidDecision("decision accept requires a claim-to-evidence relation or --judgment-only <rationale>."); }
+function invalidDecision(message: string): never { const error = new Error(message) as Error & { code: string }; error.code = "invalid_transition"; throw error; }
+
+
+export function isDecisionEvent(event: { readonly schema: string }): event is DecisionEventV1 { return event.schema === "decision-event/v1"; }
+export function serializeDecisionEvent(event: DecisionEventV1): string { const errors = validateCurrentDecisionEvent(event); if (errors.length) throw new Error(errors.join("; ")); return serializeEventEnvelope(event); }
+
+export function validateDecisionEvent(value: unknown): readonly string[] { return validateDecisionEventFields(value, true); }
+export function validateCurrentDecisionEvent(value: unknown): readonly string[] { return validateDecisionEventFields(value, false); }
+function validateDecisionEventFields(value: unknown, allowUnknownFields: boolean): readonly string[] {
+  if (!isRecord(value) || !matchesFields(value, ["schema", "eventId", "workspaceRevision", "opId", "decisionId", "type", "actor", "source", "occurredAt", "payload"], allowUnknownFields)
+    || value.schema !== "decision-event/v1" || !includes(decisionEventTypes, value.type) || !decisionId(value.decisionId) || !timestamp(value.occurredAt) || !isRecord(value.payload)) return ["decision event envelope or payload is invalid"];
+  if (validateEventEnvelopeIdentity(value, allowUnknownFields).length) return ["decision event envelope identity is invalid"];
+  const payload = value.payload, type = value.type, common = ["baseDocumentSha256", "decisionDocumentClaim"], mutation = validDecisionMutation(payload, value.decisionId, type === "decision_proposed", allowUnknownFields); if (!mutation) return ["decision document mutation is invalid"];
+  if (type === "decision_proposed") return proposalIssues(payload, common, value.decisionId, allowUnknownFields);
+  if (type === "decision_accepted") return requiredWithOptional(payload, ["rationale", "judgmentOnlyRationale", "judgmentConsent", ...common], ["fulfillments", "standingPolicy", "contentPin"], allowUnknownFields) && codePoints(payload.rationale, 1, 199) && (payload.judgmentOnlyRationale === null || codePoints(payload.judgmentOnlyRationale, 1, 199)) && (payload.fulfillments === undefined || fulfillments(payload.fulfillments, allowUnknownFields)) && (payload.standingPolicy === undefined || typeof payload.standingPolicy === "boolean") && judgmentConsent(payload.judgmentConsent, value, type, allowUnknownFields) && (payload.contentPin === undefined || contentPin(payload.contentPin, value, allowUnknownFields)) ? [] : ["decision accepted payload is invalid"];
+  if (type === "decision_rejected" || type === "decision_deferred") return requiredWithOptional(payload, ["reason", "judgmentConsent", ...common], ["contentPin"], allowUnknownFields) && codePoints(payload.reason, 1, 199) && judgmentConsent(payload.judgmentConsent, value, type, allowUnknownFields) && (payload.contentPin === undefined || contentPin(payload.contentPin, value, allowUnknownFields)) ? [] : [`${type} payload is invalid`];
+  if (type === "decision_superseded" || type === "decision_retired") return requiredWithOptional(payload, ["reason", ...common], ["contentPin"], allowUnknownFields) && codePoints(payload.reason, 1, 199) && (payload.contentPin === undefined || contentPin(payload.contentPin, value, allowUnknownFields)) ? [] : [`${type} payload is invalid`];
+  if (type === "decision_amended") return matchesFields(payload, ["next", "fields", "body", "amendment", "contentPin", ...common], allowUnknownFields) && amendableSnapshot(payload.next, allowUnknownFields) && uniqueFactStrings(payload.fields) && payload.fields.length > 0 && (payload.body === null || typeof payload.body === "string") && amendment(payload.amendment, value, allowUnknownFields) && contentPin(payload.contentPin, value, allowUnknownFields) ? [] : ["decision amendment payload is invalid"];
+  if (type === "decision_repinned") return matchesFields(payload, ["migrationEvidence", "contentPin", ...common], allowUnknownFields) && typeof payload.migrationEvidence === "string" && /^task\/[^/]+\/[^/]+$/u.test(payload.migrationEvidence) && contentPin(payload.contentPin, value, allowUnknownFields) ? [] : ["decision repin payload is invalid"];
+  if (type === "decision_claim_declared") return matchesFields(payload, ["claimId", "text", "loadBearing", ...common], allowUnknownFields) && claimId(payload.claimId) && isNonEmptyString(payload.text) && typeof payload.loadBearing === "boolean" ? [] : ["decision claim payload is invalid"];
+  if (type === "decision_claim_fulfillment_declared") return matchesFields(payload, ["claimId", "mode", ...common], allowUnknownFields) && claimId(payload.claimId) && includes(decisionFulfillmentModes, payload.mode) ? [] : ["decision fulfillment payload is invalid"];
+  if (type === "decision_related") return matchesFields(payload, ["relation", ...common], allowUnknownFields) && relation(payload.relation, allowUnknownFields) ? [] : ["decision relation payload is invalid"];
+  if (type === "decision_relation_replaced") return matchesFields(payload, ["relationId", "reason", "replacement", "body", ...common], allowUnknownFields) && relationId(payload.relationId) && codePoints(payload.reason, 1, 199) && relation(payload.replacement, allowUnknownFields) && (payload.body === null || typeof payload.body === "string") ? [] : ["decision relation replacement payload is invalid"];
+  return matchesFields(payload, ["relationId", "reason", ...common], allowUnknownFields) && relationId(payload.relationId) && codePoints(payload.reason, 1, 199) ? [] : ["decision relation retirement payload is invalid"];
+}
+
+
+function includes<const T extends readonly string[]>(values: T, value: unknown): value is T[number] { return typeof value === "string" && (values as readonly string[]).includes(value); }
+function decisionId(value: unknown): value is string { return typeof value === "string" && /^dec_[A-Za-z0-9_-]+$/u.test(value); }
+function optionId(value: unknown, prefix: "CH" | "RJ"): value is string { return typeof value === "string" && new RegExp(`^${prefix}[A-Za-z0-9_-]+$`, "u").test(value); }
+function claimId(value: unknown): value is string { return typeof value === "string" && /^C[A-Za-z0-9_-]+$/u.test(value); }
+function uniqueFactStrings(value: unknown): value is readonly string[] { return Array.isArray(value) && value.every(isNonEmptyString) && new Set(value).size === value.length; }
+// #1546: one "payload is invalid" sentence made a rejected packet a bisection exercise. Every predicate
+// below now reports the field it guards, so a single rejection names each field that is actually wrong.
+// Accept/reject semantics are unchanged: the same predicates decide, they just no longer short-circuit.
+function proposalIssues(value: Readonly<Record<string, unknown>>, common: readonly string[] = [], id?: unknown, allowUnknownFields = false): readonly string[] {
+  const fields = ["title", "question", "riskTier", "urgency", "vertical", "preset", "appliesTo", "decisionClass", "chosen", "rejected", "body", "claims", "fulfillments", "relations"];
+  if (!matchesFields(value, [...fields, ...common], allowUnknownFields)) return [`decision proposal requires exactly: ${fields.join(", ")}`];
+  const issues: string[] = [], check = (ok: boolean, message: string): void => { if (!ok) issues.push(message); };
+  check(codePoints(value.question, 1, 499), "question must be 1..499 code points");
+  for (const field of ["riskTier", "urgency"] as const) check(includes(["low", "medium", "high"] as const, value[field]), `${field} must be low, medium, or high`);
+  for (const field of ["title", "vertical", "preset"] as const) check(isNonEmptyString(value[field]), `${field} must be a non-empty string`);
+  check(includes(["ordinary", "standing_policy"] as const, value.decisionClass), "decisionClass must be ordinary or standing_policy");
+  check(isRecord(value.appliesTo) && matchesFields(value.appliesTo, ["modules", "productLines"], allowUnknownFields) && uniqueFactStrings(value.appliesTo.modules) && uniqueFactStrings(value.appliesTo.productLines), "appliesTo must carry exactly modules and productLines as arrays of unique non-empty strings");
+  check(typeof value.body === "string", "body must be a string");
+  for (const field of ["chosen", "rejected"] as const) check(Array.isArray(value[field]) && value[field].length > 0, `${field} must be a non-empty array`);
+  for (const field of ["claims", "fulfillments", "relations"] as const) check(Array.isArray(value[field]), `${field} must be an array`);
+  if (issues.length) return issues;
+  const chosen = value.chosen as readonly unknown[], rejected = value.rejected as readonly unknown[], claims = value.claims as readonly unknown[], fulfilled = value.fulfillments as readonly unknown[], relations = value.relations as readonly unknown[];
+  check(chosen.every((entry) => isRecord(entry) && requiredWithOptional(entry, ["id", "text"], ["rationale"], allowUnknownFields) && optionId(entry.id, "CH") && isNonEmptyString(entry.text) && (entry.rationale === undefined || codePoints(entry.rationale, 1, 199))), "every chosen entry needs a CH id, non-empty text, and an optional 1..199 rationale");
+  check(rejected.every((entry) => isRecord(entry) && matchesFields(entry, ["id", "text", "whyNot"], allowUnknownFields) && optionId(entry.id, "RJ") && isNonEmptyString(entry.text) && codePoints(entry.whyNot, 1, 199)), "every rejected entry needs an RJ id, non-empty text, and a 1..199 whyNot");
+  check(claims.every((entry) => isRecord(entry) && matchesFields(entry, ["id", "text", "loadBearing"], allowUnknownFields) && claimId(entry.id) && isNonEmptyString(entry.text) && typeof entry.loadBearing === "boolean"), "every claim needs a C id, non-empty text, and a boolean loadBearing");
+  check(fulfilled.every((entry) => isRecord(entry) && matchesFields(entry, ["claimId", "mode"], allowUnknownFields) && claimId(entry.claimId) && includes(decisionFulfillmentModes, entry.mode)), `every fulfillment needs a claimId and a mode of ${decisionFulfillmentModes.join(", ")}`);
+  const ids = [...chosen, ...rejected, ...claims].map((entry) => isRecord(entry) ? String(entry.id) : ""), claimIds = new Set(claims.map((entry) => isRecord(entry) ? entry.id : null)), fulfillmentIds = fulfilled.map((entry) => isRecord(entry) ? entry.claimId : null), relationIds = relations.map((entry) => isRecord(entry) ? entry.relation_id : null), anchored = new Set(ids);
+  check(new Set(ids).size === ids.length, "chosen, rejected, and claim ids must be unique across the packet");
+  check(new Set(fulfillmentIds).size === fulfillmentIds.length && fulfillmentIds.every((entry) => claimIds.has(entry)), "every fulfillment must name a distinct claim declared in this packet");
+  check(relations.every((entry) => relation(entry, allowUnknownFields) && entry.relation_id === deriveRelationId(entry) && entry.source.startsWith(`decision/${String(id)}/`) && anchored.has(entry.source.slice(entry.source.lastIndexOf("/") + 1))), "every relation must derive its own relation_id and anchor on a chosen, rejected, or claim id of this decision");
+  check(new Set(relationIds).size === relationIds.length, "relation ids must be unique");
+  return issues;
+}
+function validDecisionMutation(value: Readonly<Record<string, unknown>>, id: unknown, proposed: boolean, allowUnknownFields: boolean): boolean { const base = value.baseDocumentSha256; if (proposed ? base !== null : typeof base !== "string" || !/^[0-9a-f]{64}$/u.test(base)) return false; const claim = value.decisionDocumentClaim; if (!isRecord(claim) || !matchesFields(claim, ["path", "sha256", "size", "mediaType", "policyId"], allowUnknownFields) || claim.path !== `decisions/decision-${String(id)}/decision.md` || !/^[0-9a-f]{64}$/u.test(String(claim.sha256)) || !Number.isSafeInteger(claim.size) || (claim.size as number) < 0 || claim.mediaType !== "text/markdown" || claim.policyId !== DECISION_DOCUMENT_POLICY_ID) return false; try { return normalizeRelativeDocumentPath(String(claim.path)) === claim.path; } catch { return false; } }
+function judgmentConsent(value: unknown, event: Readonly<Record<string, unknown>>, type: DecisionOutcomeType, allowUnknownFields: boolean): value is DecisionJudgmentConsentV1 { if (!isRecord(value) || !matchesFields(value, ["schema", "consentId", "decisionId", "action", "targetState", "machineDigest", "actor", "source", "consentedAt"], allowUnknownFields)) return false; const action = type.slice("decision_".length, -2), targetState = outcomeState(type); return value.schema === "decision-judgment-consent/v1" && /^djc_[0-9a-f]{26}$/u.test(String(value.consentId)) && value.decisionId === event.decisionId && value.action === action && value.targetState === targetState && /^sha256:[0-9a-f]{64}$/u.test(String(value.machineDigest)) && validateActorIdentity(value.actor, allowUnknownFields).length === 0 && sameActorIdentity(value.actor, event.actor) && validateWriteSource(value.source, allowUnknownFields).length === 0 && sameWriteSource(value.source, event.source) && value.consentedAt === event.occurredAt; }
+function contentPin(value: unknown, event: Readonly<Record<string, unknown>>, allowUnknownFields: boolean): value is DecisionContentPinV1 { return isRecord(value) && matchesFields(value, ["schema", "pinId", "action", "state", "pinnedAt", "evidence", "actor", "digest"], allowUnknownFields) && value.schema === "decision-content-pin/v1" && /^dcp_[0-9a-f]{26}$/u.test(String(value.pinId)) && includes(["accept", "reject", "defer", "supersede", "retire", "amend", "repin"] as const, value.action) && includes(decisionStates, value.state) && value.pinnedAt === event.occurredAt && codePoints(value.evidence, 1, 199) && validateActorIdentity(value.actor, allowUnknownFields).length === 0 && sameActorIdentity(value.actor, event.actor) && /^sha256:[0-9a-f]{64}$/u.test(String(value.digest)); }
+function amendment(value: unknown, event: Readonly<Record<string, unknown>>, allowUnknownFields: boolean): value is DecisionAmendmentV1 { return isRecord(value) && matchesFields(value, ["schema", "amendmentId", "fields", "actor", "amendedAt"], allowUnknownFields) && value.schema === "decision-amendment/v1" && /^dam_[0-9a-f]{26}$/u.test(String(value.amendmentId)) && uniqueFactStrings(value.fields) && value.fields.length > 0 && validateActorIdentity(value.actor, allowUnknownFields).length === 0 && sameActorIdentity(value.actor, event.actor) && value.amendedAt === event.occurredAt; }
+function amendableSnapshot(value: unknown, allowUnknownFields: boolean): value is DecisionAmendableSnapshot { if (!isRecord(value) || !matchesFields(value, ["title", "decisionClass", "chosen", "rejected", "claims"], allowUnknownFields) || !isNonEmptyString(value.title) || !includes(["ordinary", "standing_policy"] as const, value.decisionClass) || !Array.isArray(value.chosen) || !Array.isArray(value.rejected) || !Array.isArray(value.claims)) return false; const chosen = value.chosen.every((entry) => isRecord(entry) && requiredWithOptional(entry, ["id", "text"], ["rationale"], allowUnknownFields) && optionId(entry.id, "CH") && isNonEmptyString(entry.text)), rejected = value.rejected.every((entry) => isRecord(entry) && matchesFields(entry, ["id", "text", "whyNot"], allowUnknownFields) && optionId(entry.id, "RJ") && isNonEmptyString(entry.text) && isNonEmptyString(entry.whyNot)), claims = value.claims.every((entry) => isRecord(entry) && matchesFields(entry, ["id", "text", "loadBearing", "fulfillment"], allowUnknownFields) && claimId(entry.id) && isNonEmptyString(entry.text) && typeof entry.loadBearing === "boolean" && (entry.fulfillment === null || includes(decisionFulfillmentModes, entry.fulfillment))), ids = [...value.chosen, ...value.rejected, ...value.claims].map((entry) => isRecord(entry) ? entry.id : null); return chosen && rejected && claims && new Set(ids).size === ids.length; }
+function fulfillments(value: unknown, allowUnknownFields: boolean): boolean { return Array.isArray(value) && value.every((entry) => isRecord(entry) && matchesFields(entry, ["claimId", "mode"], allowUnknownFields) && claimId(entry.claimId) && includes(decisionFulfillmentModes, entry.mode)) && new Set(value.map((entry) => isRecord(entry) ? entry.claimId : null)).size === value.length; }
+function relationId(value: unknown): value is string { return typeof value === "string" && /^rel_[0-9a-f]{16}$/u.test(value); }
+function relation(value: unknown, allowUnknownFields: boolean): value is EntityRelationRecord { return isRecord(value) && matchesFields(value, ["relation_id", "source", "target", "type", "strength", "direction", "origin", "rationale", "state"], allowUnknownFields) && typeof value.relation_id === "string" && /^rel_[0-9a-f]{16}$/u.test(value.relation_id) && isNonEmptyString(value.source) && isNonEmptyString(value.target) && includes(relationTypes, value.type) && includes(relationStrengths, value.strength) && includes(relationDirections, value.direction) && includes(relationOrigins, value.origin) && codePoints(value.rationale, 1, 199) && includes(relationStates, value.state) && value.state === "active"; }
