+++ packages/kernel/src/domain/doc-sync.contract.ts
@@ -6,7 +6,7 @@ import { canonicalizeWriteValue, freezeDeclaredWritePlan, hasOnlyFields, hasRequ
 import type { DocSyncDifference, DocSyncReceiptDetail, DocSyncUnresolvedTouch, LedgerCommitIdentity, LedgerCutIdentity, LedgerIdentity } from "./receipt-domain-registry.ts";
 import type { LeaseV1 } from "./execution.ts";
 import { validateAgentRuntimeEvent, validateCurrentAgentRuntimeEvent, type AgentRuntimeEventV1 } from "./agent-runtime.ts";
-import { isTaskBootstrapEvent, validateCurrentTaskBootstrapEvent, validateTaskBootstrapEvent, type TaskBootstrapEventV1 } from "./task-bootstrap-event.ts"; import { isDecisionEvent, isFactEvent, validateCurrentDecisionEvent, validateCurrentFactEvent, validateDecisionEvent, validateFactEvent, type DecisionEventV1, type FactEventV1 } from "./fact-event.ts";
+import { isTaskBootstrapEvent, validateCurrentTaskBootstrapEvent, validateTaskBootstrapEvent, type TaskBootstrapEventV1 } from "./task-bootstrap-event.ts"; import { isDecisionEvent, validateCurrentDecisionEvent, validateDecisionEvent, type DecisionEventV1 } from "./decision-event.ts"; import { isFactEvent, validateCurrentFactEvent, validateFactEvent, type FactEventV1 } from "./fact-event.ts";
 import { isTaskProgressEvent, validateCurrentTaskProgressEvent, validateTaskProgressEvent, type TaskProgressEventV1 } from "./task-progress-event.ts"; import { isPresetSnapshotUpgradeEvent, validateCurrentPresetSnapshotUpgradeEvent, validatePresetSnapshotUpgradeEvent, type PresetSnapshotUpgradeEventV1 } from "./preset-snapshot-upgrade-event.ts";
 import { assertNoPortablePathCollisions, normalizeRelativeDocumentPath, type PortableDocumentPath } from "../layout/portable-path.ts";
 import { isMigrationImportEvent, MIGRATION_DOCUMENT_POLICY_ID, validateCurrentMigrationImportEvent, validateMigrationImportEvent, type MigrationImportEventV1 } from "./migration-import-event.ts";
