+++ packages/kernel/src/domain/status-vocabulary.ts
@@ -216,8 +216,8 @@ export const statusVocabularies: readonly StatusVocabulary[] = [
   { id: "task.status.review-artifacts", entity: "Task", field: "status", module: "packages/kernel/src/domain/lifecycle-status.ts", anchor: "reviewArtifactStatuses", words: ["in_review", "done"], subsetOf: "task.status", note: "Statuses whose review artifacts are required." },
   { id: "task.status.wip-occupying", entity: "Task", field: "status", module: "packages/kernel/src/domain/task-wip-policy.ts", anchor: "taskWipOccupyingStatuses", words: ["active", "blocked", "in_review"], subsetOf: "task.status", note: "Statuses that occupy a WIP slot." },
   { id: "task.status.replay", entity: "Task", field: "status", module: "packages/kernel/src/domain/task.ts", anchor: "replayTaskStatuses", words: ["planned", "active", "blocked", "in_review", "done", "cancelled"], subsetOf: "task.status", note: "Task/v1 wire validator set; same words, same meanings." },
-  { id: "decision.state", entity: "Decision", field: "state", module: "packages/kernel/src/domain/fact-event.ts", anchor: "decisionStates", words: ["proposed", "in_effect", "rejected", "deferred", "superseded", "outcome_retired"] },
-  { id: "decision.judgment.target", entity: "Decision", field: "judgment targetState", module: "packages/kernel/src/domain/fact-event.ts", anchor: "#targetState", words: ["in_effect", "rejected", "deferred"], subsetOf: "decision.state", note: "Consent target states of the three judgment actions." },
+  { id: "decision.state", entity: "Decision", field: "state", module: "packages/kernel/src/domain/decision-event.ts", anchor: "decisionStates", words: ["proposed", "in_effect", "rejected", "deferred", "superseded", "outcome_retired"] },
+  { id: "decision.judgment.target", entity: "Decision", field: "judgment targetState", module: "packages/kernel/src/domain/decision-event.ts", anchor: "#targetState", words: ["in_effect", "rejected", "deferred"], subsetOf: "decision.state", note: "Consent target states of the three judgment actions." },
   { id: "execution.state", entity: "Execution", field: "state", module: "packages/kernel/src/domain/execution.ts", anchor: "executionStates", words: ["active", "submitted", "accepted", "changes_requested", "abandoned"] },
   { id: "execution.state.v1", entity: "Execution", field: "state", module: "packages/kernel/src/domain/execution.ts", anchor: "executionV1States", words: ["active", "submitted", "changes_requested", "accepted"], subsetOf: "execution.state", note: "Native execution/v1 subset (archived v0 adds abandoned)." },
   { id: "lease.phase", entity: "Lease", field: "phase", module: "packages/kernel/src/domain/execution.ts", anchor: "leasePhases", words: ["reserving", "held", "orphaned", "released"] },
