+++ packages/kernel/src/entity/field-contracts.ts
@@ -1,6 +1,7 @@
 import type { TaskFrontmatter } from "../schemas/registry.ts";
 import type { EntityRelationRecord } from "../domain/entity-relation.ts";
-import type { DecisionProposalPayload, FactEventPayload } from "../domain/fact-event.ts";
+import type { DecisionProposalPayload } from "../domain/decision-event.ts";
+import type { FactEventPayload } from "../domain/fact-event.ts";
 import { sessionFieldContracts } from "./session-declaration.ts";
 
 export type EntityKindWithFieldCoverage = "decision" | "task" | "fact" | "relation" | "session";
