+++ packages/kernel/src/index.ts
@@ -66,12 +66,14 @@ export * from "./ports/index.ts";
 export * from "./projection/post-merge-checks.ts";
 export { detectRelationGraphCycles, readRelationGraphAuthoredSourceKinds, validateRelationGraphRecords } from "./projection/relation-graph-projection.ts"; export type { FactAnchorRow, RelationCoverageRow, RelationFactRow, RelationGraphEdgeRow, RelationGraphProjection, RelationRecordEntry, RelationRecordValidationIssue } from "./projection/relation-graph-projection.ts";
 export { projectDecisionReadiness } from "./projection/decision-readiness-projection.ts";
-export type { DecisionListFilters, DecisionProjectionRow, FactProjectionRow, FactSearchFilters } from "./projection/fact-event-projection.ts";
+export type { DecisionListFilters, DecisionProjectionRow } from "./projection/decision-event-projection.ts";
+export type { FactProjectionRow, FactSearchFilters } from "./projection/fact-event-projection.ts";
 export { buildColdCoverage, readColdRebuildSource } from "./projection/cold-rebuild-source.ts";
 export type { ColdDecisionProjectionRow, ColdRebuildIssue } from "./projection/cold-rebuild-source.ts";
 export { readMarkdownSource, taskEntryToRow } from "./projection/sqlite-task-source.ts";
 export type { TaskSourceEntry } from "./projection/sqlite-task-source.ts";
-export { renderDecisionDocument, renderFactsDocument } from "./domain/fact-event.ts";
+export { renderDecisionDocument } from "./domain/decision-event.ts";
+export { renderFactsDocument } from "./domain/fact-event.ts";
 export * from "./publish/index.ts";
 export * from "./projection/sqlite-task-projection.ts";
 export { readTaskProjectionSchemaVersion, taskProjectionSchemaVersion } from "./projection/projection-schema.ts";
