+++ packages/kernel/src/projection/fact-event-projection.ts
@@ -1,16 +1,10 @@
 import type { DatabaseSync } from "node:sqlite";
-import { consumeKnownError } from "../error-consumption.ts";
-import { deriveRelationId, validateRelationRecordsForHost, type EntityRelationRecord } from "../domain/entity-relation.ts";
-import { assertDecisionContentPin, assertDecisionJudgmentConsent, decisionDocumentProse, factRef, type DecisionAmendmentV1, type DecisionContentPinV1, type DecisionDocumentState, type DecisionEventV1, type DecisionFulfillmentMode, type DecisionJudgmentConsentV1, type DecisionState, type FactEventV1, type FactMemoryClass, type FactConfidence } from "../domain/fact-event.ts";
+import { deriveRelationId } from "../domain/entity-relation.ts";
+import { factRef, type FactEventV1, type FactMemoryClass, type FactConfidence } from "../domain/fact-event.ts";
 import type { ActorIdentity, WriteSource } from "../domain/write-chain.contract.ts";
 import type { FactAnchorRow } from "./relation-graph-projection.ts";
-import type { DocumentState } from "../domain/doc-sync.contract.ts";
-import type { DecisionReadinessProjection } from "./decision-readiness-projection.ts";
-import { stableStringify } from "../integrity/stable-hash.ts";
-import { isSameExecution } from "../domain/actor-domain-services.ts";
-import { coverageOf } from "../domain/decision-coverage.ts";
 import { factLiveness } from "../domain/fact-liveness.ts";
-import { checkedPageLimit, decodePageCursor, encodePageCursor, type ProjectionPage } from "./task-query-projection.ts";
+import { ftsQuery } from "./fts-query.ts";
 
 export interface FactProjectionRow {
   readonly schema: "fact-row/v1";
@@ -37,13 +31,6 @@ export interface FactRelationEdgeRow {
 }
 export interface FactSearchFilters { readonly query?: string; readonly taskId?: string; readonly refs?: readonly string[]; readonly confidence?: FactConfidence; readonly memoryClass?: FactMemoryClass; readonly observedAfter?: string; readonly observedBefore?: string; readonly limit?: number; readonly cursor?: string }
 export interface FactSearchPage { readonly limit: number; readonly cursor: string | null; readonly nextCursor: string | null }
-export interface DecisionBodyRow { readonly path: string; readonly blobSha256: string; readonly size: number; readonly mediaType: string; readonly body: string; readonly workspaceRevision: number }
-export interface DecisionProjectionRow { readonly schema: "decision-row/v1"; readonly decisionId: string; readonly legacyId?: string; readonly path: string; readonly state: DecisionState; readonly title: string; readonly question: string; readonly riskTier: "low" | "medium" | "high"; readonly urgency: "low" | "medium" | "high"; readonly vertical: string; readonly preset: string; readonly decisionClass: "ordinary" | "standing_policy"; readonly appliesTo: { readonly modules: readonly string[]; readonly productLines: readonly string[] }; readonly proposer: ActorIdentity; readonly arbiter: ActorIdentity | null; readonly proposedAt: string; readonly decidedAt: string | null; readonly workspaceRevision: number; readonly chosen: readonly { readonly id: string; readonly text: string; readonly rationale?: string }[]; readonly rejected: readonly { readonly id: string; readonly text: string; readonly whyNot: string }[]; readonly claims: readonly { readonly id: string; readonly text: string; readonly loadBearing: boolean; readonly fulfillment: DecisionFulfillmentMode | null }[]; readonly judgmentConsents: readonly DecisionJudgmentConsentV1[]; readonly amendments?: readonly DecisionAmendmentV1[]; readonly contentPins?: readonly DecisionContentPinV1[]; readonly body: DecisionBodyRow | null; readonly readiness?: DecisionReadinessProjection }
-export type DecisionAgendaProjectionRow = Pick<DecisionProjectionRow, "decisionId" | "title" | "riskTier" | "urgency" | "proposedAt">; export interface DecisionPageQuery { readonly state: DecisionState; readonly limit?: number; readonly cursor?: string }
-export interface DecisionListFilters { readonly search?: string; readonly legacyId?: string; readonly legacyRange?: { readonly start: number; readonly end: number }; readonly state?: DecisionState; readonly module?: string; readonly productLine?: string }
-export interface DecisionAnchorRow { readonly decisionRef: string; readonly decisionId: string; readonly anchorRefs: readonly string[]; readonly sourcePath: string }
-export interface DecisionRelationEdgeRow { readonly relationId: string; readonly sourceRef: string; readonly targetRef: string; readonly relationType: EntityRelationRecord["type"]; readonly direction: EntityRelationRecord["direction"]; readonly strength: EntityRelationRecord["strength"]; readonly origin: EntityRelationRecord["origin"]; readonly state: "active" | "edge_retired"; readonly rationale: string; readonly ownerRef: string; readonly sourcePath: string; readonly recordIndex: number; readonly retiredRevision?: number; readonly retiredAt?: string; readonly retirementReason?: string }
-export interface DecisionCoverageRow { readonly decisionRef: string; readonly claimRef: string; readonly status: "covered" | "uncovered"; readonly fulfillment: DecisionFulfillmentMode | null; readonly coveringFactRef?: string; readonly refutingFactRefs: readonly string[]; readonly relationPath: readonly string[]; readonly basisRevision: number }
 export class FactProjectionError extends Error { readonly code: "content_not_ready" | "invalid_transition" | "entity_not_found" | "anchor_not_found" | "relation_invalid";
   constructor(code: FactProjectionError["code"], message: string) { super(message); this.name = "FactProjectionError"; this.code = code; } }
 
@@ -52,107 +39,14 @@ export function createFactProjectionTables(db: DatabaseSync): void {
     CREATE TABLE IF NOT EXISTS fact (task_id TEXT NOT NULL, fact_id TEXT NOT NULL, ref TEXT NOT NULL UNIQUE, statement TEXT NOT NULL, evidence_source TEXT NOT NULL,
       observed_at TEXT NOT NULL, confidence TEXT NOT NULL, memory_class TEXT NOT NULL, op_id TEXT NOT NULL UNIQUE, workspace_revision INTEGER NOT NULL, row_json TEXT NOT NULL,
       PRIMARY KEY(task_id, fact_id));
-    CREATE TABLE IF NOT EXISTS relation_edge (relation_id TEXT PRIMARY KEY, source_ref TEXT NOT NULL, target_ref TEXT NOT NULL, relation_type TEXT NOT NULL,
-      state TEXT NOT NULL, owner_ref TEXT NOT NULL, workspace_revision INTEGER NOT NULL, row_json TEXT NOT NULL);
     CREATE VIRTUAL TABLE IF NOT EXISTS fact_fts USING fts5(task_id UNINDEXED, fact_id UNINDEXED, statement, evidence_source,
       tokenize='unicode61 remove_diacritics 2');
     CREATE INDEX IF NOT EXISTS fact_filter ON fact(task_id, confidence, memory_class, observed_at);
     CREATE INDEX IF NOT EXISTS fact_observed_page ON fact(observed_at DESC, task_id ASC, fact_id ASC);
     CREATE INDEX IF NOT EXISTS fact_confidence_page ON fact(confidence, observed_at DESC, task_id ASC, fact_id ASC);
     CREATE INDEX IF NOT EXISTS fact_memory_class_page ON fact(memory_class, observed_at DESC, task_id ASC, fact_id ASC);
-    CREATE INDEX IF NOT EXISTS relation_edge_source ON relation_edge(source_ref, state);
-    CREATE INDEX IF NOT EXISTS relation_edge_target ON relation_edge(target_ref, state);
-    CREATE INDEX IF NOT EXISTS relation_edge_type_target ON relation_edge(relation_type, target_ref, state);
-    CREATE INDEX IF NOT EXISTS relation_edge_state_page ON relation_edge(state, relation_id);
-    CREATE TABLE IF NOT EXISTS decision (decision_id TEXT PRIMARY KEY, state TEXT NOT NULL, title TEXT NOT NULL, question TEXT NOT NULL, risk_tier TEXT NOT NULL, urgency TEXT NOT NULL, vertical TEXT NOT NULL, preset TEXT NOT NULL, decision_class TEXT NOT NULL, applies_json TEXT NOT NULL, proposer_json TEXT NOT NULL, arbiter_json TEXT, proposed_at TEXT NOT NULL, decided_at TEXT, workspace_revision INTEGER NOT NULL);
-    CREATE TABLE IF NOT EXISTS decision_option (decision_id TEXT NOT NULL, kind TEXT NOT NULL, option_id TEXT NOT NULL, position INTEGER NOT NULL, text TEXT NOT NULL, rationale TEXT, workspace_revision INTEGER NOT NULL, PRIMARY KEY(decision_id, kind, option_id));
-    CREATE TABLE IF NOT EXISTS decision_claim (decision_id TEXT NOT NULL, claim_id TEXT NOT NULL, position INTEGER NOT NULL, text TEXT NOT NULL, load_bearing INTEGER NOT NULL, fulfillment TEXT, declared_revision INTEGER NOT NULL, fulfilled_revision INTEGER, PRIMARY KEY(decision_id, claim_id));
-    CREATE TABLE IF NOT EXISTS decision_judgment_consent (consent_id TEXT PRIMARY KEY, decision_id TEXT NOT NULL, workspace_revision INTEGER NOT NULL, value_json TEXT NOT NULL);
-    CREATE TABLE IF NOT EXISTS decision_amendment (amendment_id TEXT PRIMARY KEY, decision_id TEXT NOT NULL, workspace_revision INTEGER NOT NULL, value_json TEXT NOT NULL);
-    CREATE TABLE IF NOT EXISTS decision_content_pin (pin_id TEXT PRIMARY KEY, decision_id TEXT NOT NULL, workspace_revision INTEGER NOT NULL, value_json TEXT NOT NULL);
-    CREATE VIRTUAL TABLE IF NOT EXISTS decision_fts USING fts5(decision_id UNINDEXED, title, question, option_text, claim_text, body, tokenize='unicode61 remove_diacritics 2');
-    CREATE INDEX IF NOT EXISTS decision_filter ON decision(state, risk_tier, urgency, vertical);
-    CREATE INDEX IF NOT EXISTS decision_state_page ON decision(state, decision_id);
-    CREATE INDEX IF NOT EXISTS decision_judgment_consent_owner ON decision_judgment_consent(decision_id, workspace_revision);
-    CREATE INDEX IF NOT EXISTS decision_amendment_owner ON decision_amendment(decision_id, workspace_revision);
-    CREATE INDEX IF NOT EXISTS decision_content_pin_owner ON decision_content_pin(decision_id, workspace_revision);
   `);
 }
-
-export function assertDecisionAdmission(db: DatabaseSync, event: DecisionEventV1): void { const row = decisionState(db, event.decisionId); if (event.type === "decision_proposed") { if (row) fail("invalid_transition", `Decision ${event.decisionId} already exists.`); const host = `decision/${event.decisionId}`, anchors = new Set([...event.payload.chosen, ...event.payload.rejected, ...event.payload.claims].map((entry) => `${host}/${entry.id}`)); if (validateRelationRecordsForHost(host, event.payload.relations).length) fail("relation_invalid", "Initial relation owner, kind triple, rationale, or deterministic id is invalid."); for (const relation of event.payload.relations) { if (!anchors.has(relation.source)) fail("anchor_not_found", `Relation source ${relation.source} is not owned by this Decision.`); if (relation.target !== host && !anchors.has(relation.target) && !knownEndpoint(db, relation.target)) fail("entity_not_found", `Relation target ${relation.target} does not exist.`); if (db.prepare("SELECT 1 FROM relation_edge WHERE relation_id=?").get(relation.relation_id)) fail("relation_invalid", `Relation ${relation.relation_id} already exists.`); } return; } if (!row) fail("entity_not_found", `Decision ${event.decisionId} does not exist.`);
-  if (["decision_accepted", "decision_rejected", "decision_deferred"].includes(event.type)) { if (row.state !== "proposed") fail("invalid_transition", `${event.type} requires proposed state.`); if (event.actor.executor !== null && isSameExecution(JSON.parse(row.proposer_json) as ActorIdentity, event.actor)) fail("invalid_transition", "An agent cannot judge its own Decision proposal."); const current = readDecisionDocumentState(db, event.decisionId); if (!current) fail("entity_not_found", `Decision ${event.decisionId} does not exist.`); try { assertDecisionJudgmentConsent(current, event as Extract<DecisionEventV1, { readonly type: "decision_accepted" | "decision_rejected" | "decision_deferred" }>); assertDecisionContentPin(current, event); } catch (error) { consumeKnownError(error); fail("invalid_transition", error instanceof Error ? error.message : String(error)); } if (event.type === "decision_accepted") { for (const fulfillment of event.payload.fulfillments ?? []) { const claim = current.claims.find((entry) => entry.id === fulfillment.claimId); if (!claim) fail("anchor_not_found", `Claim ${fulfillment.claimId} does not exist.`); if (claim.fulfillment) fail("invalid_transition", `Claim ${fulfillment.claimId} already has a fulfillment.`); } if (event.payload.standingPolicy && current.appliesTo.modules.length + current.appliesTo.productLines.length === 0) fail("invalid_transition", "Standing policy acceptance requires a non-empty applies_to scope."); } return; }
-  if (event.type === "decision_superseded" || event.type === "decision_retired") { if (row.state !== "in_effect") fail("invalid_transition", `${event.type} requires in_effect state.`); const current = readDecisionDocumentState(db, event.decisionId)!; try { assertDecisionContentPin(current, event); } catch (error) { consumeKnownError(error); fail("invalid_transition", error instanceof Error ? error.message : String(error)); } return; }
-  if (event.type === "decision_amended") { assertAmendment(db, event); const current = readDecisionDocumentState(db, event.decisionId)!; try { assertDecisionContentPin(current, event); } catch (error) { consumeKnownError(error); fail("invalid_transition", error instanceof Error ? error.message : String(error)); } return; }
-  if (event.type === "decision_repinned") { const current = readDecisionDocumentState(db, event.decisionId)!; try { assertDecisionContentPin(current, event); } catch (error) { consumeKnownError(error); fail("invalid_transition", error instanceof Error ? error.message : String(error)); } return; }
-  if (event.type === "decision_claim_declared") { if (!(["proposed", "in_effect"] as const).includes(row.state as never)) fail("invalid_transition", "Claims require proposed or in_effect state."); if (db.prepare("SELECT 1 FROM decision_claim WHERE decision_id=? AND claim_id=?").get(event.decisionId, event.payload.claimId)) fail("invalid_transition", `Claim ${event.payload.claimId} already exists.`); return; }
-  if (event.type === "decision_claim_fulfillment_declared") { const claim = db.prepare("SELECT fulfillment FROM decision_claim WHERE decision_id=? AND claim_id=?").get(event.decisionId, event.payload.claimId) as { readonly fulfillment: string | null } | undefined; if (!claim) fail("anchor_not_found", `Claim ${event.payload.claimId} does not exist.`); if (claim.fulfillment) fail("invalid_transition", `Claim ${event.payload.claimId} already has a fulfillment.`); return; }
-  if (event.type === "decision_related") { const relation = event.payload.relation, host = `decision/${event.decisionId}`; if (validateRelationRecordsForHost(host, [relation]).length) fail("relation_invalid", "Relation owner, kind triple, rationale, or deterministic id is invalid."); if (!knownDecisionSource(db, event.decisionId, relation.source)) fail("anchor_not_found", `Relation source ${relation.source} is not owned by this Decision.`); if (!knownEndpoint(db, relation.target)) fail("entity_not_found", `Relation target ${relation.target} does not exist.`); if (db.prepare("SELECT 1 FROM relation_edge WHERE relation_id=?").get(relation.relation_id)) fail("relation_invalid", `Relation ${relation.relation_id} already exists.`); return; }
-  if (event.type !== "decision_relation_retired" && event.type !== "decision_relation_replaced") fail("invalid_transition", "Unsupported Decision event."); const edge = db.prepare("SELECT state, owner_ref FROM relation_edge WHERE relation_id=?").get(event.payload.relationId) as { readonly state: string; readonly owner_ref: string } | undefined; if (!edge) fail("entity_not_found", `Relation ${event.payload.relationId} does not exist.`); if (edge.owner_ref !== `decision/${event.decisionId}` || edge.state !== "active") fail("relation_invalid", `Relation ${event.payload.relationId} is not an active edge owned by this Decision.`); if (event.type === "decision_relation_replaced") { const relation = event.payload.replacement, host = `decision/${event.decisionId}`; if (validateRelationRecordsForHost(host, [relation]).length) fail("relation_invalid", "Replacement relation owner, kind triple, rationale, or deterministic id is invalid."); if (!knownDecisionSource(db, event.decisionId, relation.source)) fail("anchor_not_found", `Relation source ${relation.source} is not owned by this Decision.`); if (!knownEndpoint(db, relation.target)) fail("entity_not_found", `Relation target ${relation.target} does not exist.`); if (db.prepare("SELECT 1 FROM relation_edge WHERE relation_id=?").get(relation.relation_id)) fail("relation_invalid", `Replacement relation ${relation.relation_id} already exists.`); }
-}
-
-export function reduceDecisionEvent(db: DatabaseSync, event: DecisionEventV1): void { assertDecisionAdmission(db, event); const revision = event.workspaceRevision;
-  if (event.type === "decision_proposed") { const p = event.payload; db.prepare("INSERT INTO decision VALUES (?, 'proposed', ?, ?, ?, ?, ?, ?, ?, ?, ?, NULL, ?, NULL, ?)").run(event.decisionId, p.title, p.question, p.riskTier, p.urgency, p.vertical, p.preset, p.decisionClass, JSON.stringify(p.appliesTo), JSON.stringify(event.actor), event.occurredAt, revision); const insert = db.prepare("INSERT INTO decision_option VALUES (?, ?, ?, ?, ?, ?, ?)"); for (const [position, option] of p.chosen.entries()) insert.run(event.decisionId, "chosen", option.id, position, option.text, option.rationale ?? null, revision); for (const [position, option] of p.rejected.entries()) insert.run(event.decisionId, "rejected", option.id, position, option.text, option.whyNot, revision); const claim = db.prepare("INSERT INTO decision_claim VALUES (?, ?, ?, ?, ?, ?, ?, ?)"); for (const [position, entry] of p.claims.entries()) { const fulfillment = p.fulfillments.find((candidate) => candidate.claimId === entry.id)?.mode ?? null; claim.run(event.decisionId, entry.id, position, entry.text, entry.loadBearing ? 1 : 0, fulfillment, revision, fulfillment ? revision : null); } for (const [index, relation] of p.relations.entries()) insertDecisionRelation(db, event, relation, index); refreshDecisionFts(db, event.decisionId); return; }
-  db.prepare("UPDATE decision SET workspace_revision=? WHERE decision_id=?").run(revision, event.decisionId);
-  if (event.type === "decision_accepted" || event.type === "decision_rejected" || event.type === "decision_deferred") { const state = event.type === "decision_accepted" ? "in_effect" : event.type.slice("decision_".length); db.prepare("UPDATE decision SET state=?, arbiter_json=?, decided_at=?, workspace_revision=? WHERE decision_id=?").run(state, JSON.stringify(event.actor), event.occurredAt, revision, event.decisionId); db.prepare("INSERT INTO decision_judgment_consent VALUES (?, ?, ?, ?)").run(event.payload.judgmentConsent.consentId, event.decisionId, revision, JSON.stringify(event.payload.judgmentConsent)); if (event.type === "decision_accepted") { if (event.payload.standingPolicy) db.prepare("UPDATE decision SET decision_class='standing_policy' WHERE decision_id=?").run(event.decisionId); for (const fulfillment of event.payload.fulfillments ?? []) db.prepare("UPDATE decision_claim SET fulfillment=?, fulfilled_revision=? WHERE decision_id=? AND claim_id=?").run(fulfillment.mode, revision, event.decisionId, fulfillment.claimId); } insertDecisionPin(db, event); refreshDecisionFts(db, event.decisionId); return; }
-  if (event.type === "decision_superseded" || event.type === "decision_retired") { const state = event.type === "decision_superseded" ? "superseded" : "outcome_retired"; db.prepare("UPDATE decision SET state=?, decided_at=?, workspace_revision=? WHERE decision_id=?").run(state, event.occurredAt, revision, event.decisionId); insertDecisionPin(db, event); refreshDecisionFts(db, event.decisionId); return; }
-  if (event.type === "decision_amended") { const next = event.payload.next; db.prepare("UPDATE decision SET title=?, decision_class=? WHERE decision_id=?").run(next.title, next.decisionClass, event.decisionId); db.prepare("DELETE FROM decision_option WHERE decision_id=?").run(event.decisionId); const option = db.prepare("INSERT INTO decision_option VALUES (?, ?, ?, ?, ?, ?, ?)"); for (const [position, value] of next.chosen.entries()) option.run(event.decisionId, "chosen", value.id, position, value.text, value.rationale ?? null, revision); for (const [position, value] of next.rejected.entries()) option.run(event.decisionId, "rejected", value.id, position, value.text, value.whyNot, revision); db.prepare("DELETE FROM decision_claim WHERE decision_id=?").run(event.decisionId); const claim = db.prepare("INSERT INTO decision_claim VALUES (?, ?, ?, ?, ?, ?, ?, ?)"); for (const [position, value] of next.claims.entries()) claim.run(event.decisionId, value.id, position, value.text, value.loadBearing ? 1 : 0, value.fulfillment, revision, value.fulfillment ? revision : null); db.prepare("INSERT INTO decision_amendment VALUES (?, ?, ?, ?)").run(event.payload.amendment.amendmentId, event.decisionId, revision, JSON.stringify(event.payload.amendment)); insertDecisionPin(db, event); refreshDecisionFts(db, event.decisionId); return; }
-  if (event.type === "decision_repinned") { insertDecisionPin(db, event); return; }
-  if (event.type === "decision_claim_declared") { db.prepare("INSERT INTO decision_claim VALUES (?, ?, (SELECT COALESCE(MAX(position), -1) + 1 FROM decision_claim WHERE decision_id=?), ?, ?, NULL, ?, NULL)").run(event.decisionId, event.payload.claimId, event.decisionId, event.payload.text, event.payload.loadBearing ? 1 : 0, revision); refreshDecisionFts(db, event.decisionId); return; }
-  if (event.type === "decision_claim_fulfillment_declared") { db.prepare("UPDATE decision_claim SET fulfillment=?, fulfilled_revision=? WHERE decision_id=? AND claim_id=?").run(event.payload.mode, revision, event.decisionId, event.payload.claimId); return; }
-  if (event.type === "decision_related") { insertDecisionRelation(db, event, event.payload.relation, 0); return; }
-  if (event.type !== "decision_relation_retired" && event.type !== "decision_relation_replaced") fail("invalid_transition", "Unsupported Decision event."); retireDecisionRelation(db, event, event.payload.relationId, event.payload.reason); if (event.type === "decision_relation_replaced") insertDecisionRelation(db, event, event.payload.replacement, 0);
-}
-function insertDecisionRelation(db: DatabaseSync, event: DecisionEventV1, relation: EntityRelationRecord, recordIndex: number): void { const edge: DecisionRelationEdgeRow = { relationId: relation.relation_id, sourceRef: relation.source, targetRef: relation.target, relationType: relation.type, direction: relation.direction, strength: relation.strength, origin: relation.origin, state: "active", rationale: relation.rationale, ownerRef: `decision/${event.decisionId}`, sourcePath: `event:${event.opId}`, recordIndex }; db.prepare("INSERT INTO relation_edge VALUES (?, ?, ?, ?, ?, ?, ?, ?)").run(edge.relationId, edge.sourceRef, edge.targetRef, edge.relationType, edge.state, edge.ownerRef, event.workspaceRevision, JSON.stringify(edge)); }
-function retireDecisionRelation(db: DatabaseSync, event: DecisionEventV1, relationId: string, reason: string): void { const current = JSON.parse(String((db.prepare("SELECT row_json FROM relation_edge WHERE relation_id=?").get(relationId) as { readonly row_json: string }).row_json)) as DecisionRelationEdgeRow, retired = { ...current, state: "edge_retired" as const, retiredRevision: event.workspaceRevision, retiredAt: event.occurredAt, retirementReason: reason }; db.prepare("UPDATE relation_edge SET state='edge_retired', workspace_revision=?, row_json=? WHERE relation_id=?").run(event.workspaceRevision, JSON.stringify(retired), relationId); }
-function insertDecisionPin(db: DatabaseSync, event: DecisionEventV1): void { if (!("contentPin" in event.payload) || event.payload.contentPin === undefined) return; const pin = event.payload.contentPin; db.prepare("INSERT INTO decision_content_pin VALUES (?, ?, ?, ?)").run(pin.pinId, event.decisionId, event.workspaceRevision, JSON.stringify(pin)); }
-
-export function refreshDecisionDocumentSearch(db: DatabaseSync, document: DocumentState): void { const id = /^decisions\/decision-(dec_[A-Za-z0-9_-]+)\/decision\.md$/u.exec(document.path)?.[1]; if (id && decisionState(db, id)) refreshDecisionFts(db, id); }
-export function readDecisionRow(db: DatabaseSync, decisionId: string, withBody = true): DecisionProjectionRow | null { return readDecisionRows(db, [decisionId], withBody)[0] ?? null; }
-export function readDecisionRows(db: DatabaseSync, decisionIds: readonly string[], withBody = true): readonly DecisionProjectionRow[] {
-  if (!Array.isArray(decisionIds) || decisionIds.some((decisionId) => typeof decisionId !== "string" || decisionId.length === 0)) throw new Error("decision collection read requires non-empty string ids");
-  if (decisionIds.length === 0) return [];
-  const bodyColumn = withBody ? "document.value_json" : "NULL", bodyJoin = withBody ? "LEFT JOIN document ON document.path = 'decisions/decision-' || decision.decision_id || '/decision.md'" : "";
-  const sql = `WITH requested_decisions(request_order, decision_id) AS MATERIALIZED (SELECT CAST(key AS INTEGER), value FROM json_each(?))
-    SELECT requested_decisions.request_order, decision.decision_id, decision.state, decision.title, decision.question, decision.risk_tier, decision.urgency, decision.vertical, decision.preset, decision.decision_class, decision.applies_json, decision.proposer_json, decision.arbiter_json, decision.proposed_at, decision.decided_at, decision.workspace_revision,
-      ${bodyColumn} AS body_document_json,
-      COALESCE((SELECT json_group_array(json_object('kind', kind, 'option_id', option_id, 'text', text, 'rationale', rationale)) FROM (SELECT kind, option_id, text, rationale FROM decision_option WHERE decision_id = decision.decision_id ORDER BY kind, position)), '[]') AS options_json,
-      COALESCE((SELECT json_group_array(json_object('claim_id', claim_id, 'text', text, 'load_bearing', load_bearing, 'fulfillment', fulfillment)) FROM (SELECT claim_id, text, load_bearing, fulfillment FROM decision_claim WHERE decision_id = decision.decision_id ORDER BY position)), '[]') AS claims_json,
-      COALESCE((SELECT json_group_array(value_json) FROM (SELECT value_json FROM decision_judgment_consent WHERE decision_id = decision.decision_id ORDER BY workspace_revision)), '[]') AS consents_json,
-      COALESCE((SELECT json_group_array(value_json) FROM (SELECT value_json FROM decision_amendment WHERE decision_id = decision.decision_id ORDER BY workspace_revision)), '[]') AS amendments_json,
-      COALESCE((SELECT json_group_array(value_json) FROM (SELECT value_json FROM decision_content_pin WHERE decision_id = decision.decision_id ORDER BY workspace_revision)), '[]') AS pins_json
-    FROM requested_decisions JOIN decision ON decision.decision_id = requested_decisions.decision_id ${bodyJoin}
-    ORDER BY requested_decisions.request_order`;
-  return (db.prepare(sql).all(JSON.stringify(decisionIds)) as unknown as readonly DecisionCollectionRecord[]).map(decisionCollectionRow);
-}
-interface DecisionCollectionRecord extends Record<string, unknown> { readonly decision_id: string; readonly state: string; readonly title: string; readonly question: string; readonly risk_tier: string; readonly urgency: string; readonly vertical: string; readonly preset: string; readonly decision_class: string; readonly applies_json: string; readonly proposer_json: string; readonly arbiter_json: string | null; readonly proposed_at: string; readonly decided_at: string | null; readonly workspace_revision: number; readonly body_document_json: string | null; readonly options_json: string; readonly claims_json: string; readonly consents_json: string; readonly amendments_json: string; readonly pins_json: string }
-function decisionCollectionRow(row: DecisionCollectionRecord): DecisionProjectionRow {
-  const decisionId = row.decision_id, legacyId = decisionLegacyId(decisionId), options = JSON.parse(row.options_json) as { readonly kind: string; readonly option_id: string; readonly text: string; readonly rationale: string | null }[], claims = JSON.parse(row.claims_json) as { readonly claim_id: string; readonly text: string; readonly load_bearing: number; readonly fulfillment: DecisionFulfillmentMode | null }[], consents = (JSON.parse(row.consents_json) as string[]).map((value) => JSON.parse(value) as DecisionJudgmentConsentV1), amendments = (JSON.parse(row.amendments_json) as string[]).map((value) => JSON.parse(value) as DecisionAmendmentV1), pins = (JSON.parse(row.pins_json) as string[]).map((value) => JSON.parse(value) as DecisionContentPinV1), body = row.body_document_json === null ? null : decisionBodyFromDocument(decisionId, row.body_document_json);
-  return { schema: "decision-row/v1", decisionId, ...(legacyId ? { legacyId } : {}), path: `decisions/decision-${decisionId}/decision.md`, state: row.state as DecisionState, title: row.title, question: row.question, riskTier: row.risk_tier as DecisionProjectionRow["riskTier"], urgency: row.urgency as DecisionProjectionRow["urgency"], vertical: row.vertical, preset: row.preset, decisionClass: row.decision_class as DecisionProjectionRow["decisionClass"], appliesTo: JSON.parse(row.applies_json) as DecisionProjectionRow["appliesTo"], proposer: JSON.parse(row.proposer_json) as ActorIdentity, arbiter: row.arbiter_json === null ? null : JSON.parse(row.arbiter_json) as ActorIdentity, proposedAt: row.proposed_at, decidedAt: row.decided_at, workspaceRevision: Number(row.workspace_revision), chosen: options.filter((option) => option.kind === "chosen").map((option) => ({ id: option.option_id, text: option.text, ...(option.rationale ? { rationale: option.rationale } : {}) })), rejected: options.filter((option) => option.kind === "rejected").map((option) => ({ id: option.option_id, text: option.text, whyNot: option.rationale! })), claims: claims.map((claim) => ({ id: claim.claim_id, text: claim.text, loadBearing: Boolean(claim.load_bearing), fulfillment: claim.fulfillment })), judgmentConsents: consents, ...(amendments.length ? { amendments } : {}), ...(pins.length ? { contentPins: pins } : {}), body };
-}
-export function listDecisionRows(db: DatabaseSync, filters: DecisionListFilters): readonly DecisionProjectionRow[] { const where: string[] = [], values: string[] = []; if (filters.search?.trim()) { where.push("decision_id IN (SELECT decision_id FROM decision_fts WHERE decision_fts MATCH ?)"); values.push(ftsQuery(filters.search)); } if (filters.state) { where.push("state=?"); values.push(filters.state); } if (filters.module) { where.push("EXISTS (SELECT 1 FROM json_each(decision.applies_json, '$.modules') WHERE value=?)"); values.push(filters.module); } if (filters.productLine) { where.push("EXISTS (SELECT 1 FROM json_each(decision.applies_json, '$.productLines') WHERE value=?)"); values.push(filters.productLine); } return (db.prepare(`SELECT decision_id FROM decision${where.length ? ` WHERE ${where.join(" AND ")}` : ""}`).all(...values) as unknown as readonly { readonly decision_id: string }[]).map((row) => row.decision_id).filter((decisionId) => { const legacy = decisionLegacyId(decisionId); return (!filters.legacyId || legacy === filters.legacyId) && (!filters.legacyRange || legacy !== undefined && projectedLegacyNumber(legacy) >= filters.legacyRange.start && projectedLegacyNumber(legacy) <= filters.legacyRange.end); }).sort(compareDecisionIds).map((decisionId) => ({ ...readDecisionRow(db, decisionId, false)!, body: null })); }
-export function listDecisionAgendaRowsPage(db: DatabaseSync, query: DecisionPageQuery): { readonly rows: readonly DecisionAgendaProjectionRow[]; readonly page: ProjectionPage } { const values: (string | number)[] = [query.state], where = ["state = ?"]; if (query.cursor !== undefined) { const [decisionId] = decodePageCursor(query.cursor, 1); where.push("decision_id > ?"); values.push(decisionId!); } const limit = query.limit === undefined ? 100 : checkedPageLimit(query.limit); values.push(limit + 1); const raw = db.prepare(`SELECT decision_id, title, risk_tier, urgency, proposed_at FROM decision WHERE ${where.join(" AND ")} ORDER BY decision_id LIMIT ?`).all(...values) as unknown as readonly { readonly decision_id: string; readonly title: string; readonly risk_tier: DecisionAgendaProjectionRow["riskTier"]; readonly urgency: DecisionAgendaProjectionRow["urgency"]; readonly proposed_at: string }[], visible = raw.slice(0, limit), rows = visible.map((row) => ({ decisionId: row.decision_id, title: row.title, riskTier: row.risk_tier, urgency: row.urgency, proposedAt: row.proposed_at })), last = visible.at(-1); return { rows, page: { limit, cursor: query.cursor ?? null, nextCursor: raw.length > limit && last ? encodePageCursor([last.decision_id]) : null } }; }
-export function decisionLegacyId(decisionId: string): string | undefined { const match = /(?:^|_)E([1-9][0-9]*)(?:_|$)/u.exec(decisionId); return match ? `E${Number(match[1])}` : undefined; }
-function projectedLegacyNumber(value: string): number { return Number(value.slice(1)); } function compareDecisionIds(left: string, right: string): number { const legacyLeft = decisionLegacyId(left), legacyRight = decisionLegacyId(right), a = legacyLeft ? projectedLegacyNumber(legacyLeft) : undefined, b = legacyRight ? projectedLegacyNumber(legacyRight) : undefined; return a !== undefined && b !== undefined && a !== b ? a - b : a !== undefined && b === undefined ? -1 : a === undefined && b !== undefined ? 1 : left.localeCompare(right); }
-export function readDecisionGraphRows(db: DatabaseSync): { readonly edges: readonly DecisionRelationEdgeRow[]; readonly decisionAnchors: readonly DecisionAnchorRow[]; readonly coverageRows: readonly DecisionCoverageRow[] } { const edges = (db.prepare("SELECT row_json FROM relation_edge WHERE owner_ref NOT LIKE 'fact/%' ORDER BY relation_id").all() as unknown as readonly { readonly row_json: string }[]).map((r) => JSON.parse(r.row_json) as DecisionRelationEdgeRow), anchors = decisionAnchorIndex(db), decisionAnchors = (db.prepare("SELECT decision_id FROM decision ORDER BY decision_id").all() as unknown as readonly { readonly decision_id: string }[]).map((r) => ({ decisionId: r.decision_id, decisionRef: `decision/${r.decision_id}`, anchorRefs: anchors.get(r.decision_id) ?? [`decision/${r.decision_id}`], sourcePath: `event:decision/${r.decision_id}` })); return { edges, decisionAnchors, coverageRows: decisionCoverage(db, edges) }; }
-function decisionAnchorIndex(db: DatabaseSync): ReadonlyMap<string, readonly string[]> { const map = new Map<string, string[]>(); for (const row of db.prepare("SELECT decision_id FROM decision").all() as unknown as readonly { readonly decision_id: string }[]) map.set(row.decision_id, [`decision/${row.decision_id}`]); for (const sql of ["SELECT decision_id, option_id AS anchor FROM decision_option", "SELECT decision_id, claim_id AS anchor FROM decision_claim"]) for (const row of db.prepare(sql).all() as unknown as readonly { readonly decision_id: string; readonly anchor: string }[]) map.get(row.decision_id)?.push(`decision/${row.decision_id}/${row.anchor}`); for (const refs of map.values()) refs.sort(); return map; }
-
-function decisionCoverage(db: DatabaseSync, edges: readonly DecisionRelationEdgeRow[]): readonly DecisionCoverageRow[] {
-  const basisRevision = Number((db.prepare("SELECT watermark FROM projection_meta WHERE singleton=1").get() as { readonly watermark: number }).watermark), decisions = db.prepare("SELECT decision_id,state,decision_class,applies_json FROM decision ORDER BY decision_id").all() as unknown as readonly { readonly decision_id: string; readonly state: string; readonly decision_class: string; readonly applies_json: string }[], claims = db.prepare("SELECT decision_id,claim_id,load_bearing,fulfillment FROM decision_claim ORDER BY decision_id,claim_id").all() as unknown as readonly { readonly decision_id: string; readonly claim_id: string; readonly load_bearing: number; readonly fulfillment: DecisionFulfillmentMode | null }[], facts = db.prepare("SELECT ref FROM fact ORDER BY ref").all() as unknown as readonly { readonly ref: string }[], hasTasks = Boolean(db.prepare("SELECT 1 FROM sqlite_master WHERE type='table' AND name='task_snapshot'").get()), tasks = hasTasks ? db.prepare("SELECT task_id,json_extract(snapshot_json,'$.task.status') AS status FROM task_snapshot ORDER BY task_id").all() as unknown as readonly { readonly task_id: string; readonly status: string }[] : [];
-  const claimsByDecision = new Map<string, typeof claims>();
-  for (const claim of claims) claimsByDecision.set(claim.decision_id, [...claimsByDecision.get(claim.decision_id) ?? [], claim]);
-  const livenessEdges = db.prepare("SELECT relation_id AS relationId, source_ref AS sourceRef, target_ref AS targetRef, relation_type AS relationType, state FROM relation_edge WHERE relation_type='supersedes-fact' ORDER BY relation_id").all() as unknown as readonly { readonly relationId: string; readonly sourceRef: string; readonly targetRef: string; readonly relationType: string; readonly state: string }[];
-  return coverageOf(decisions.map((decision) => { const ref = `decision/${decision.decision_id}`; return { ref, state: decision.state, decisionClass: decision.decision_class, appliesTo: JSON.parse(decision.applies_json) as { readonly modules: readonly string[]; readonly productLines: readonly string[] }, claims: (claimsByDecision.get(decision.decision_id) ?? []).map((claim) => ({ ref: `${ref}/${claim.claim_id}`, loadBearing: claim.load_bearing === 1, fulfillment: claim.fulfillment })) }; }), facts, tasks.map((task) => ({ ref: `task/${task.task_id}`, status: task.status })), [...edges, ...livenessEdges]).map((row) => ({ ...row, fulfillment: row.fulfillment === "standing-policy" ? "standing_policy" : row.fulfillment, basisRevision }));
-}
-function refreshDecisionFts(db: DatabaseSync, decisionId: string): void { const row = db.prepare("SELECT title,question FROM decision WHERE decision_id=?").get(decisionId) as { readonly title: string; readonly question: string } | undefined; if (!row) return; const options = (db.prepare("SELECT text FROM decision_option WHERE decision_id=?").all(decisionId) as unknown as readonly { readonly text: string }[]).map((o) => o.text).join(" "), claims = (db.prepare("SELECT text FROM decision_claim WHERE decision_id=?").all(decisionId) as unknown as readonly { readonly text: string }[]).map((c) => c.text).join(" "), body = readDecisionBody(db, decisionId)?.body ?? ""; db.prepare("DELETE FROM decision_fts WHERE decision_id=?").run(decisionId); db.prepare("INSERT INTO decision_fts VALUES (?, ?, ?, ?, ?, ?)").run(decisionId, row.title, row.question, options, claims, body); }
-function readDecisionBody(db: DatabaseSync, decisionId: string): DecisionBodyRow | null { const path = `decisions/decision-${decisionId}/decision.md`, row = db.prepare("SELECT value_json FROM document WHERE path=?").get(path) as { readonly value_json: string } | undefined; return row ? decisionBodyFromDocument(decisionId, row.value_json) : null; }
-function decisionBodyFromDocument(decisionId: string, valueJson: string): DecisionBodyRow { const path = `decisions/decision-${decisionId}/decision.md`, document = JSON.parse(valueJson) as DocumentState; return { path, blobSha256: document.blobSha256, size: document.size, mediaType: document.mediaType, body: decisionDocumentProse(document.body), workspaceRevision: document.workspaceRevision }; }
-export function readDecisionDocumentState(db: DatabaseSync, decisionId: string): DecisionDocumentState | null { const row = readDecisionRow(db, decisionId); if (!row) return null; const { body: _body, ...decision } = row, relations = (db.prepare("SELECT row_json FROM relation_edge WHERE owner_ref=? ORDER BY relation_id").all(`decision/${decisionId}`) as unknown as readonly { readonly row_json: string }[]).map((entry) => JSON.parse(entry.row_json) as DecisionRelationEdgeRow).map((edge) => ({ relation_id: edge.relationId, source: edge.sourceRef, target: edge.targetRef, type: edge.relationType, direction: edge.direction, strength: edge.strength, origin: edge.origin, rationale: edge.rationale, state: edge.state })); return { ...decision, relations }; }
-function assertAmendment(db: DatabaseSync, event: Extract<DecisionEventV1, { readonly type: "decision_amended" }>): void { const current = readDecisionDocumentState(db, event.decisionId); if (!current) fail("entity_not_found", `Decision ${event.decisionId} does not exist.`); const next = event.payload.next, fields = new Set(event.payload.fields), changed = new Set<string>(), prefix = <T>(before: readonly T[], after: readonly T[]) => after.length >= before.length && before.every((value, index) => stableStringify(value) === stableStringify(after[index])); if (next.title !== current.title) changed.add("title"); if (next.decisionClass !== current.decisionClass) changed.add("decisionClass"); if (!prefix(current.chosen, next.chosen)) fail("invalid_transition", "Amendment may append chosen options but cannot rewrite or remove them."); if (next.chosen.length !== current.chosen.length) changed.add("chosen"); if (!prefix(current.rejected, next.rejected)) fail("invalid_transition", "Amendment may append rejected options but cannot rewrite or remove them."); if (next.rejected.length !== current.rejected.length) changed.add("rejected"); if (next.claims.length < current.claims.length) fail("invalid_transition", "Amendment cannot remove claims."); for (const [index, claim] of current.claims.entries()) { const replacement = next.claims[index]; if (!replacement || replacement.id !== claim.id || replacement.text !== claim.text) fail("invalid_transition", "Amendment may append claims but cannot rewrite their identities or text."); if (claim.fulfillment !== null && replacement.fulfillment !== claim.fulfillment) fail("invalid_transition", "Amendment cannot change an existing claim fulfillment."); } if (stableStringify(next.claims) !== stableStringify(current.claims)) changed.add("claims"); if (event.payload.body !== null) changed.add("body"); if ([...fields].some((field) => !["title", "decisionClass", "chosen", "rejected", "claims", "body"].includes(field)) || [...fields].sort().join("\0") !== [...changed].sort().join("\0")) fail("invalid_transition", `Amendment fields must exactly name the changed machine and prose channels (declared: ${[...fields].join(",")}; changed: ${[...changed].join(",")}).`); if (JSON.stringify(event.payload.amendment.fields) !== JSON.stringify(event.payload.fields)) fail("invalid_transition", "Amendment history does not match the declared field set."); }
-function decisionState(db: DatabaseSync, id: string): { readonly state: DecisionState; readonly proposer_json: string } | undefined { return db.prepare("SELECT state,proposer_json FROM decision WHERE decision_id=?").get(id) as { readonly state: DecisionState; readonly proposer_json: string } | undefined; }
-function decisionAnchorRefs(db: DatabaseSync, id: string): string[] { const root = `decision/${id}`, option = db.prepare("SELECT option_id FROM decision_option WHERE decision_id=?").all(id) as unknown as readonly { readonly option_id: string }[], claims = db.prepare("SELECT claim_id FROM decision_claim WHERE decision_id=?").all(id) as unknown as readonly { readonly claim_id: string }[]; return [root, ...option.map((o) => `${root}/${o.option_id}`), ...claims.map((c) => `${root}/${c.claim_id}`)].sort(); }
-function knownDecisionSource(db: DatabaseSync, id: string, ref: string): boolean { return decisionAnchorRefs(db, id).includes(ref); }
-function knownEndpoint(db: DatabaseSync, ref: string): boolean { const decision = /^decision\/([^/]+)(?:\/([^/]+))?$/u.exec(ref); if (decision) return decision[2] ? decisionAnchorRefs(db, decision[1]!).includes(ref) : Boolean(decisionState(db, decision[1]!)); const fact = /^fact\/([^/]+)\/(F-[0-9A-HJKMNP-TV-Z]{8})$/u.exec(ref); if (fact) return Boolean(db.prepare("SELECT 1 FROM fact WHERE task_id=? AND fact_id=?").get(fact[1], fact[2])); const task = /^task\/([^/]+)$/u.exec(ref); return Boolean(task && db.prepare("SELECT 1 FROM task_snapshot WHERE task_id=?").get(task[1])); }
-function fail(code: FactProjectionError["code"], message: string): never { throw new FactProjectionError(code, message); }
 
 export function assertFactAdmission(db: DatabaseSync, event: FactEventV1): void {
   const ownRef = factRef(event.taskId, event.factId);
@@ -238,7 +132,6 @@ export function readFactGraphRows(db: DatabaseSync): { readonly edges: readonly 
 }
 
 export function readFactAnchorRows(db: DatabaseSync, refs?: readonly string[]): readonly FactAnchorRow[] { if (refs !== undefined && refs.length === 0) return []; const scoped = refs !== undefined && refs.length <= 900; const where = scoped ? ` WHERE ref IN (${refs.map(() => "?").join(",")})` : ""; const memoryRefs = refs !== undefined && !scoped ? new Set(refs) : null; const rows = (db.prepare(`SELECT ref, task_id, fact_id, op_id FROM fact${where} ORDER BY ref`).all(...(scoped ? refs : [])) as unknown as readonly { readonly ref: string; readonly task_id: string; readonly fact_id: string; readonly op_id: string }[]).map((row) => ({ factRef: row.ref, taskId: row.task_id, factId: row.fact_id, sourcePath: `event:${row.op_id}` })); return memoryRefs === null ? rows : rows.filter((row) => memoryRefs.has(row.factRef)); }
-function ftsQuery(value: string): string { return `"${value.trim().replaceAll('"', '""')}"`; }
 function checkedFactPageLimit(value: number): number { if (!Number.isSafeInteger(value) || value < 1 || value > 500) throw new Error("fact query page limit must be an integer between 1 and 500"); return value; }
 function encodeFactCursor(parts: readonly string[]): string { return Buffer.from(JSON.stringify(parts), "utf8").toString("base64url"); }
 function decodeFactCursor(value: string): readonly [string, string] { let parsed: unknown; try { parsed = JSON.parse(Buffer.from(value, "base64url").toString("utf8")); } catch { throw new Error("fact query cursor is invalid"); } if (!Array.isArray(parsed) || parsed.length !== 2 || parsed.some((part) => typeof part !== "string" || part.length === 0)) throw new Error("fact query cursor is invalid"); return [parsed[0] as string, parsed[1] as string]; }
