+++ packages/kernel/src/projection/fts-query.ts
@@ -1,0 +1,1 @@
+export function ftsQuery(value: string): string { return `"${value.trim().replaceAll('"', '""')}"`; }
