+++ packages/kernel/src/schemas/registry.ts
@@ -5,7 +5,8 @@ import { priorityTiers, taskWorkKinds } from "../domain/task-metadata.ts";
 import type { LifecycleBinding } from "../domain/lifecycle-binding.ts";
 import { LinkKindSchema, ProvenanceEntrySchema } from "./common.ts";
 import { EntityRelationsSchema } from "./entity-relations.ts";
-import { DecisionEventSchema, FactEventSchema } from "./fact-event.ts";
+import { DecisionEventSchema } from "./decision-event.ts";
+import { FactEventSchema } from "./fact-event.ts";
 import { HarnessCheckReportSchema } from "./harness-check-report.ts";
 import { SubtaskPlanSchema } from "./subtask-plan.ts";
 import { VerticalDefinitionSchema } from "./vertical-definition.ts";
