+++ packages/kernel/src/store/task-event-store.ts
@@ -6,7 +6,8 @@ import { assertAgentEntityWritePlan, isAgentEntityEvent } from "../domain/agent-
 import { assertDocSyncWritePlan, isDecisionEvent, isDocEvent, isFactEvent, isMigrationImportEvent, isTaskEvent, ledgerCommitSha, parseCanonicalEvent, serializeCanonicalEvent, validateCurrentCanonicalEvent, validateCurrentDocEvent, type CanonicalEventV1, type DocEventV1, type LedgerCommitSha } from "../domain/doc-sync.contract.ts";
 import { assertMigrationImportWritePlan, migrationImportClaims, migrationImportContentClaims } from "../domain/migration-import-event.ts";
 import { assertLedgerLayoutMigrationWritePlan, isLedgerLayoutMigrationEvent, ledgerLayoutMigrationWritePlan, type LedgerLayoutMigrationEventV1 } from "../domain/ledger-layout-migration-event.ts";
-import { assertDecisionWritePlan, assertFactWritePlan } from "../domain/fact-event.ts";
+import { assertDecisionWritePlan } from "../domain/decision-event.ts";
+import { assertFactWritePlan } from "../domain/fact-event.ts";
 import type { TaskEventV1 } from "../domain/task-lifecycle.contract.ts";
 import { assertTaskLifecycleWritePlan } from "../domain/task-lifecycle-publication.ts";
 import { assertTaskBootstrapWritePlan, isTaskBootstrapEvent, taskBootstrapClaims, type TaskBootstrapEventV1 } from "../domain/task-bootstrap-event.ts";
