+++ packages/kernel/test/contracts/schema-frontmatter.test.ts
@@ -4,7 +4,7 @@ import { readFile } from "node:fs/promises";
 import test from "node:test";
 import { Schema } from "effect";
 import { EntityRelationsSchema, TaskFrontmatterSchema } from "../../src/schemas/registry.ts";
-import { DecisionEventSchema } from "../../src/schemas/fact-event.ts";
+import { DecisionEventSchema } from "../../src/schemas/decision-event.ts";
 
 const validFixtureUrl = new URL("../../fixtures/schemas/task-frontmatter/valid.json", import.meta.url);
 const validDecisionFixtureUrl = new URL("../../fixtures/schemas/decision-package/valid.json", import.meta.url);
