+++ packages/kernel/test/store/relation-graph-projection.test.ts
@@ -6,7 +6,9 @@ import path from "node:path";
 import { DatabaseSync } from "node:sqlite";
 import test from "node:test";
 import { MIGRATION_DOCUMENT_POLICY_ID, REPLAY_TASK_GRAPH, checkTaskProjection, compileDecisionWrite, compileFactWrite, decisionWritePlan, deriveRelationId, formatRelationFlowRecord, makeTaskProjection, projectDecisionReadiness, readRelationGraphProjection, rebuildTaskProjection, renderDecisionDocument, serializeCanonicalEvent, sha256Text, taskLifecycleWritePlan, type DecisionEventDraftV1, type EntityRelationRecord, type FactEventDraftV1, type MigrationImportEventV1, type TaskEventV1 } from "../../src/index.ts";
-import { createFactProjectionTables, readDecisionDocumentState, reduceDecisionEvent } from "../../src/projection/fact-event-projection.ts";
+import { createDecisionProjectionTables, readDecisionDocumentState, reduceDecisionEvent } from "../../src/projection/decision-event-projection.ts";
+import { createFactProjectionTables } from "../../src/projection/fact-event-projection.ts";
+import { createRelationGraphProjectionTables } from "../../src/projection/relation-graph-projection.ts";
 import { withTempStore } from "./helpers.ts";
 
 const actor = { principal: { personId: "proposer" }, executor: null } as const;
@@ -56,7 +58,7 @@ function migrationRelationEvent(revision: number, record: EntityRelationRecord):
 function writeMigrationEvent(rootDir: string, event: MigrationImportEventV1): void { const eventRoot = path.join(rootDir, "harness/events/fixture"); mkdirSync(eventRoot, { recursive: true }); writeFileSync(path.join(eventRoot, `${event.opId}.json`), serializeCanonicalEvent(event)); }
 function writeFactEvent(rootDir: string, event: FactEventDraftV1): void { const compiled = compileFactWrite({ event, packagePath: `tasks/${event.taskId}-fixture`, currentFacts: [] }), eventRoot = path.join(rootDir, "harness/events/fixture"); mkdirSync(eventRoot, { recursive: true }); writeFileSync(path.join(eventRoot, `${event.opId}.json`), serializeCanonicalEvent(compiled.event)); }
 function projectionFixture(rootDir: string) { const blobs = new Map<string, Uint8Array>(), eventStore = { readHead: () => null, readBatch: () => ({ sourceRevision: 0, events: [], cursor: null, done: true, accessedItems: 0 }), readContentBlob: (sha256: string) => blobs.get(sha256) ?? null }; return { blobs, projection: makeTaskProjection({ rootDir, eventStore }) }; }
-function decisionProjectionDatabase(): DatabaseSync { const db = new DatabaseSync(":memory:"); db.exec("CREATE TABLE projection_meta (singleton INTEGER PRIMARY KEY, watermark INTEGER NOT NULL); INSERT INTO projection_meta VALUES (1, 1); CREATE TABLE document (path TEXT PRIMARY KEY, workspace_revision INTEGER NOT NULL, value_json TEXT NOT NULL); CREATE TABLE task_snapshot (task_id TEXT PRIMARY KEY, workspace_revision INTEGER NOT NULL, snapshot_json TEXT NOT NULL)"); createFactProjectionTables(db); return db; }
+function decisionProjectionDatabase(): DatabaseSync { const db = new DatabaseSync(":memory:"); db.exec("CREATE TABLE projection_meta (singleton INTEGER PRIMARY KEY, watermark INTEGER NOT NULL); INSERT INTO projection_meta VALUES (1, 1); CREATE TABLE document (path TEXT PRIMARY KEY, workspace_revision INTEGER NOT NULL, value_json TEXT NOT NULL); CREATE TABLE task_snapshot (task_id TEXT PRIMARY KEY, workspace_revision INTEGER NOT NULL, snapshot_json TEXT NOT NULL)"); createRelationGraphProjectionTables(db); createFactProjectionTables(db); createDecisionProjectionTables(db); return db; }
 function compileCurrent(fixture: ReturnType<typeof projectionFixture>, event: DecisionEventDraftV1) { const { projection } = fixture, current = projection.readDecision(event.decisionId).decision, path = `decisions/decision-${event.decisionId}/decision.md`, document = projection.readDocument(path).document, relations = projection.readDecisionGraph().edges.filter((edge) => edge.ownerRef === `decision/${event.decisionId}`).map((edge) => ({ relation_id: edge.relationId, source: edge.sourceRef, target: edge.targetRef, type: edge.relationType, strength: edge.strength, direction: edge.direction, origin: edge.origin, rationale: edge.rationale, state: edge.state })); return compileDecisionWrite({ event, currentDecision: current, currentRelations: relations, currentDocument: document }); }
 function applyDecision(fixture: ReturnType<typeof projectionFixture>, event: DecisionEventDraftV1): void { const compiled = compileCurrent(fixture, event); fixture.blobs.set(compiled.blobs[0].sha256, Buffer.from(compiled.body)); fixture.projection.apply(compiled.event, compiled.plan); }
 function applyFact(fixture: ReturnType<typeof projectionFixture>, event: FactEventDraftV1): void { const compiled = compileFactWrite({ event, packagePath: `tasks/${event.taskId}-evidence`, currentFacts: fixture.projection.searchFacts({ taskId: event.taskId }).facts }); fixture.blobs.set(compiled.blobs[0].sha256, Buffer.from(compiled.body)); fixture.projection.apply(compiled.event, compiled.plan); }
