+++ tools/gates/test/fixtures/decision-fact-slice-a-production-paths.json
@@ -1,12 +1,17 @@
 [
-  { "path": "packages/kernel/src/domain/fact-event.ts", "module": "decision-fact" },
-  { "path": "packages/kernel/src/schemas/fact-event.ts", "module": "decision-fact" },
-  { "path": "packages/kernel/src/projection/fact-event-projection.ts", "module": "decision-fact" },
-  { "path": "packages/application/src/fact-service.ts", "module": "decision-fact" },
-  { "path": "packages/daemon/src/fact-actions.ts", "module": "decision-fact" },
-  { "path": "packages/cli/src/cli/fact-command.ts", "module": "decision-fact" },
+  { "path": "packages/kernel/src/domain/decision-event.ts", "module": "decision" },
+  { "path": "packages/kernel/src/schemas/decision-event.ts", "module": "decision" },
+  { "path": "packages/kernel/src/projection/decision-event-projection.ts", "module": "decision" },
+  { "path": "packages/application/src/decision-service.ts", "module": "decision" },
+  { "path": "packages/daemon/src/decision-actions.ts", "module": "decision" },
+  { "path": "packages/kernel/src/domain/fact-event.ts", "module": "fact" },
+  { "path": "packages/kernel/src/schemas/fact-event.ts", "module": "fact" },
+  { "path": "packages/kernel/src/projection/fact-event-projection.ts", "module": "fact" },
+  { "path": "packages/application/src/fact-service.ts", "module": "fact" },
+  { "path": "packages/daemon/src/fact-actions.ts", "module": "fact" },
   { "path": "packages/kernel/src/domain/doc-sync.contract.ts", "module": "doc-sync" },
   { "path": "packages/kernel/src/projection/rebuildable-task-projection.ts", "module": "kernel" },
+  { "path": "packages/kernel/src/projection/relation-graph-projection.ts", "module": "kernel" },
   { "path": "packages/daemon/src/protocol/daemon-protocol.contract.ts", "module": "daemon" },
   { "path": "packages/daemon/src/repo-cell.ts", "module": "daemon" },
   { "path": "packages/cli/src/cli/thin-command.ts", "module": "cli" },
