+++ tools/gates/test/fixtures/migration-import-production-paths.json
@@ -15,7 +15,7 @@
   { "path": "packages/kernel/src/domain/write-chain.contract.ts", "module": "write-contract" },
   { "path": "packages/kernel/src/index.ts", "module": "kernel" },
   { "path": "packages/kernel/src/projection/cold-rebuild-source.ts", "module": "kernel" },
-  { "path": "packages/kernel/src/projection/fact-event-projection.ts", "module": "decision-fact" },
+  { "path": "packages/kernel/src/projection/fact-event-projection.ts", "module": "fact" },
   { "path": "packages/kernel/src/projection/rebuildable-task-projection.ts", "module": "kernel" },
   { "path": "packages/kernel/src/store/task-event-store.ts", "module": "kernel" }
 ]
