+++ tools/gates/test/fixtures/w2b-s1-production-paths.json
@@ -1,16 +1,16 @@
 [
-  { "path": "packages/application/src/fact-service.ts", "module": "decision-fact" },
-  { "path": "packages/daemon/src/fact-actions.ts", "module": "decision-fact" },
+  { "path": "packages/application/src/fact-service.ts", "module": "fact" },
+  { "path": "packages/daemon/src/fact-actions.ts", "module": "fact" },
   { "path": "packages/daemon/src/protocol/daemon-protocol.contract.ts", "module": "daemon" },
   { "path": "packages/daemon/src/repo-cell.ts", "module": "daemon" },
   { "path": "packages/gui/src/renderer/triadic-data.ts", "module": "gui" },
   { "path": "packages/kernel/src/composition/index.ts", "module": "kernel" },
-  { "path": "packages/kernel/src/domain/fact-event.ts", "module": "decision-fact" },
+  { "path": "packages/kernel/src/domain/fact-event.ts", "module": "fact" },
   { "path": "packages/kernel/src/domain/index.ts", "module": "kernel" },
   { "path": "packages/kernel/src/entity/execution-declaration.ts", "module": "kernel" },
   { "path": "packages/kernel/src/index.ts", "module": "kernel" },
   { "path": "packages/kernel/src/layout/index.ts", "module": "kernel" },
-  { "path": "packages/kernel/src/projection/fact-event-projection.ts", "module": "decision-fact" },
+  { "path": "packages/kernel/src/projection/fact-event-projection.ts", "module": "fact" },
   { "path": "packages/kernel/src/projection/rebuildable-task-projection.ts", "module": "kernel" },
   { "path": "packages/kernel/src/projection/sqlite-task-projection.ts", "module": "kernel" },
   { "path": "packages/kernel/src/store/task-event-store.ts", "module": "kernel" }
