+++ tools/gates/test/fixtures/w2b-s3-production-paths.json
@@ -1,10 +1,10 @@
 [
   { "path": "packages/cli/src/cli/thin-command.ts", "module": "cli" },
-  { "path": "packages/daemon/src/fact-actions.ts", "module": "decision-fact" },
+  { "path": "packages/daemon/src/fact-actions.ts", "module": "fact" },
   { "path": "packages/daemon/src/protocol/daemon-protocol.contract.ts", "module": "daemon" },
   { "path": "packages/daemon/src/repo-cell.ts", "module": "daemon" },
-  { "path": "packages/kernel/src/domain/fact-event.ts", "module": "decision-fact" },
+  { "path": "packages/kernel/src/domain/fact-event.ts", "module": "fact" },
   { "path": "packages/kernel/src/entity/field-contracts.ts", "module": "kernel" },
-  { "path": "packages/kernel/src/projection/fact-event-projection.ts", "module": "decision-fact" },
+  { "path": "packages/kernel/src/projection/fact-event-projection.ts", "module": "fact" },
   { "path": "packages/kernel/src/projection/rebuildable-task-projection.ts", "module": "kernel" }
 ]
